% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adhoc_table}
\alias{adhoc_table}
\title{Dispacher function for making table}
\usage{
adhoc_table(
  result_vpa,
  yrs_preabc,
  number,
  sbtarget = NULL,
  fmsy = NULL,
  data_future = NULL,
  yr_biopar = NULL,
  result_msy = NULL,
  yr_latest = NULL
)
}
\arguments{
\item{yrs_preabc}{Vector of future years preceding ABC year}

\item{number}{Table number}

\item{sbtarget}{Value of SB target}

\item{fmsy}{Value of Fmsy}
}
\description{
Dispacher function for making table
}
