% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{apply_year_colum}
\alias{apply_year_colum}
\title{列が年である行列に対して、年を指定するとその年のあいだの平均値（等）を返す関数}
\usage{
apply_year_colum(mat, target_year, stat = "mean")
}
\description{
列が年である行列に対して、年を指定するとその年のあいだの平均値（等）を返す関数
}
