% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\name{arrange_weight}
\alias{arrange_weight}
\title{モデルのAICから計算される重みに応じて、複数のモデルからシミュレーション結果を抜き出す}
\usage{
arrange_weight(weight, nsim)
}
\arguments{
\item{weight}{複数モデルのAkaike weightのベクトル}

\item{nsim}{シミュレーション回数}
}
\description{
モデルのAICから計算される重みに応じて、複数のモデルからシミュレーション結果を抜き出す
}
