% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{autocalc_ridgevpa}
\alias{autocalc_ridgevpa}
\title{リッジVPAのλ探索の自動化}
\usage{
autocalc_ridgevpa(input, target_retro, n_retro = 5, b_fix = TRUE, bin = 0.1)
}
\arguments{
\item{input}{`vpa`関数の引数をlist形式で与える}

\item{target_retro}{ペナルティλの選択に用いるmohn's rhoのパラメータ}

\item{n_retro}{レトロスペクティブ解析で遡る年数。デフォルトは`5`}

\item{b_fix}{レトロスペクティブ解析内でbを固定するか。デフォルトは`TRUE`}
}
\description{
リッジVPAのλ探索の自動化
}
\details{
etaのオプションに対応できているかは怪しい。
}
