% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{autocor.plot}
\alias{autocor.plot}
\title{再生産関係における残差の時系列の自己相関等についてのプロット}
\usage{
autocor.plot(
  resSR,
  use.resid = 1,
  lag.max = NULL,
  output = FALSE,
  filename = "Residual_trend",
  pch = 16,
  lwd = 2,
  cex = 1.2,
  cex.main = 1.2,
  cex.lab = 1.2,
  ...
)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{use.resid}{再生産関係との残差 (deviance \code{resid}) を使うか (1:default)、自己相関を除いた残差 (\code{resid2}) を使うか (2)}

\item{output}{pngファイルに出力するか否か}
}
\description{
1) 残差のトレンド、2) \code{acf}関数による自己相関係数のプロット、3) Ljung-Box検定におけるP値の3つの図を出力
}
