% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r, R/future_tmb.r
\encoding{UTF-8}
\name{average_SR_mat}
\alias{average_SR_mat}
\title{モデル平均的な再生産関係を与える}
\usage{
average_SR_mat(
  res_vpa,
  res_SR_list,
  range_list = list(1:500, 501:1000),
  SR_mat,
  seed_number,
  start_random_rec_year_name,
  recruit_age,
  resid_type = "lognormal",
  resample_year_range = 0,
  recruit_intercept = recruit_intercept,
  bias_correction = TRUE
)

average_SR_mat(
  res_vpa,
  res_SR_list,
  range_list = list(1:500, 501:1000),
  SR_mat,
  seed_number,
  start_random_rec_year_name,
  recruit_age,
  resid_type = "lognormal",
  resample_year_range = 0,
  recruit_intercept = recruit_intercept,
  bias_correction = TRUE
)
}
\arguments{
\item{res_vpa}{VPAの推定結果}

\item{res_SR_list}{再生産関係の推定結果のリスト}

\item{range_list}{}

\item{SR_mat}{将来予測用の再生産関係パラメータを格納する３次元行列}

\item{seed_number}{シード番号}

\item{start_random_rec_year_name}{ランダム加入を仮定する最初の年}

\item{resid_type}{残差の発生パターン；対数正規分布は"lognormal"、単純リサンプリングは"resampling"}

\item{resample_year_range}{0の場合、推定に使ったデータから計算される残差を用いる。年の範囲を入れると、対象とした年の範囲で計算される残差を用いる}
}
\description{
モデル平均的な再生産関係を与える

モデル平均的な再生産関係を与える
}
