% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{BH.recAR}
\alias{BH.recAR}
\title{BHを仮定したときの加入関数}
\usage{
BH.recAR(ssb, vpares, deterministic = FALSE, rec.resample = NULL,
  rec.arg = list(a = 1000, b = 1000, sd = 0.1, bias.correction = TRUE))
}
\arguments{
\item{ssb}{親魚資源量}

\item{vpares}{VPAの出力結果}
}
\description{
BHを仮定したときの加入関数
}
