% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\name{boot_steepness}
\alias{boot_steepness}
\title{再生産関係の推定結果からbootstrapを実行し、hの分布を計算する関数}
\usage{
boot_steepness(res_SR, M, waa, maa, n = 100, plus_group = TRUE)
}
\arguments{
\item{res_SR}{fit.SRまたはfit.SRregime、モデル平均の結果}

\item{M}{年齢別自然死亡係数 (ベクトルで与えるか、年齢共通の場合\code{M=0.4}のようにしてもよい)}

\item{waa}{（親魚量の）年齢別体重}

\item{maa}{年齢別親魚量}

\item{plus_group}{最高齢がプラスグループかどうか}
}
\description{
再生産関係の推定結果からbootstrapを実行し、hの分布を計算する関数
}
