% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{bootSR.ggplot}
\alias{bootSR.ggplot}
\title{再生産関係のブートストラップの結果をggplotで生成する関数する関数}
\usage{
bootSR.ggplot(boot.res, CI = 0.8)
}
\arguments{
\item{boot.res}{\code{boot.SR}のオブジェクト}

\item{CI}{プロットする信頼区間}
}
\description{
再生産関係のブートストラップの結果をggplotで生成する関数する関数
}
