% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{bootSR.plot}
\alias{bootSR.plot}
\title{再生産関係の残差ブートストラップをプロットする関数}
\usage{
bootSR.plot(
  boot.res,
  CI = 0.8,
  output = FALSE,
  filename = "boot",
  lwd = 1.2,
  pch = 1,
  ggplt = TRUE,
  ...
)
}
\arguments{
\item{boot.res}{\code{boot.SR}のオブジェクト}

\item{CI}{プロットする信頼区間}

\item{output}{pngファイルに出力するか否か}

\item{filename}{ファイル名}
}
\description{
ブートストラップによる各パラメータのヒストグラムと、それらパラメータをつかった再生産関係が描画される
}
\details{
ヒストグラムの描画ではとその中央値、上側・下側CIが（デフォルトのggplt=Tでは加えて平均値が）推定値とともに表示される

ヒストグラムのプロットにはggplotをつかうオプションggplt=TRUEをデフォルトにしているが、ggplotを使わない作図もggplt=Fで実行可能
　
ggplotで描画する場合でRstudio利用時にはplotをZoomで表示にしないと描画されないことがあるので注意
}
