% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{bubble_plot}
\alias{bubble_plot}
\title{半期VPA専用のバブルプロット作図の関数}
\usage{
bubble_plot(
  res,
  target = "naa",
  season = 1,
  hanki = TRUE,
  years = NULL,
  scale = 100,
  fix_ratio = 2,
  range = seq(0, 100, len = 5),
  digits = 0,
  max_size = 10
)
}
\description{
半期VPA専用のバブルプロット作図の関数
}
