% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{calc.rel.abund}
\alias{calc.rel.abund}
\title{年齢別パラメータを与えて、１年分前進計算する}
\usage{
calc.rel.abund(
  sel,
  Fr,
  na,
  M,
  waa,
  waa.catch = NULL,
  maa,
  min.age = 0,
  max.age = Inf,
  Pope = TRUE,
  ssb.coef = 0,
  p.pope = 0.5
)
}
\arguments{
\item{min.age}{最小の年齢（０歳がデフォルト）。数でなくて名前}

\item{max.age}{最大の年齢。数でなくて名前なので、０からスタートして２歳までの場合には、min.age=0, max.age=2とするか、min.age=1, max.age=3とする}
}
\description{
年齢別パラメータを与えて、１年分前進計算する
}
