% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{calc.residAR}
\alias{calc.residAR}
\title{再生産関係の残差から事後的に1次の自己相関係数を推定する関数}
\usage{
calc.residAR(resSR, per_regime = TRUE)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{per_regime}{自己相関係数をレジームごとに推定するか (\code{TRUE}) 否か}
}
\description{
計算は\code{fit.SR}の\code{out.AR}と同じであるが、AICcも見れるほか、\code{fit.SRregime}にも対応している
\code{fit.SRregime}の場合、各レジームの最初の年は初期値となる（つまり前のレジームの最後の年からの残差を引きずらない）
}
