% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{calc.StdResid}
\alias{calc.StdResid}
\title{再生産関係の推定における標準化残差を計算する関数}
\usage{
calc.StdResid(resSR)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}
}
\description{
再生産関係の推定における標準化残差を計算する関数
}
