% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{calc_kobeII_matrix}
\alias{calc_kobeII_matrix}
\title{Kobe II matrixを計算するための関数}
\usage{
calc_kobeII_matrix(fres_base, refs_base, Btarget = c("Btarget0"),
  Blimit = c("Blimit0"), Bban = c("Bban0"), year.lag = 0,
  beta = seq(from = 0.5, to = 1, by = 0.1))
}
\arguments{
\item{fres_base}{future.vpaの結果のオブジェクト}

\item{refs_base}{est.MSYから得られる管理基準値の表}
}
\description{
Kobe II matrixを計算するための関数
}
