% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{calc_steepness}
\alias{calc_steepness}
\title{Steepness (h) と関連するパラメータ (SB0,R0,B0)を計算する関数}
\usage{
calc_steepness(
  SR = "HS",
  rec_pars,
  M,
  waa,
  maa,
  plus_group = TRUE,
  faa = NULL,
  Pope = TRUE,
  p.pope = 0.5
)
}
\arguments{
\item{SR}{"HS", "BH", "RI"のいずれか}

\item{rec_pars}{再生産関係のパラメータで\code{rec_pars$a},\code{rec_pars$b}で使用する}

\item{M}{年齢別自然死亡係数 (ベクトルで与えるか、年齢共通の場合\code{M=0.4}のようにしてもよい)}

\item{waa}{（親魚量の）年齢別体重}

\item{maa}{年齢別親魚量}

\item{plus_group}{最高齢がプラスグループかどうか}

\item{faa}{選択率（年齢の長さのベクトル）を与えると、その選択率のもとでの決定論的なMSY管理基準値（SPRmsy,SBmsy,Rmsy,Bmsy,MSY,Fmsy2F) も返す}

\item{Pope}{faaを与えたときのYield計算にPopeの式を使うかどうか}

\item{p.pope}{faaを与えたときのYield計算でPopeの式でどこで漁獲するか（0.5=年の真ん中）}
}
\value{
以下の要素からなるデータフレーム
\describe{
\item{\code{SPR0}}{F=0のときのSPR(この逆数がreplacement lineの傾き)}
\item{\code{SB0}}{F=0のときの親魚量}
\item{\code{R0}}{F=0のときの加入量}
\item{\code{B0}}{F=0のときの資源量}
\item{\code{h}}{steepness: BHかRIのときは0.2×SB0のときの加入量がh×R0, HSのときはh=1-b/SB0}
}
}
\description{
Steepness (h) と関連するパラメータ (SB0,R0,B0)を計算する関数
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
               out.AR = TRUE)
rec_pars = resSR$pars
year <- "2017"
M = res_vpa$input$dat$M[,year]
waa = res_vpa$input$dat$waa[,year]
maa = res_vpa$input$dat$maa[,year]
Res_h = calc_steepness(SR="HS",rec_pars=rec_pars,M=M,waa=waa,maa=maa,plus_group=TRUE)
Res_h
}
}
