% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{check.SRdist}
\alias{check.SRdist}
\title{再生産関係の残差の確率分布に関するチェック図を出力する関数}
\usage{
check.SRdist(resSR, test.ks = TRUE, output = FALSE, filename = "SR_error_dist")
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{output}{pngファイルに出力するか否か}
}
\description{
1) 正規性をチェックするための検定結果・ヒストグラム、2) 累積確率分布のヒストグラム、3) 一様分布を使用した累積確率分布のQQプロttの3つが出力される
標準化した残差を使用する
累積確率分布はL2の場合は正規分布、L1の場合はラプラス分布を使用する
自己相関の外側推定の場合は、2段階で推定しているため、ファイルが2つ出力される
}
