% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{check.SRfit}
\alias{check.SRfit}
\title{再生産関係推定が収束しているかや最適解を得られているかを診断する関数}
\usage{
check.SRfit(
  resSR,
  n = 100,
  sigma = 5,
  seed = 1,
  output = FALSE,
  filename = "checkSRfit",
  fun_when_check5_replace = median
)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{n}{初期値を変えてパラメータ推定する回数}

\item{sigma}{初期値を変えるときの生起乱数の標準偏差}

\item{seed}{\code{set.seed}で使用するseed}

\item{output}{テキストファイルに結果を出力するか}

\item{filename}{ファイル名('.txt')がつく}

\item{fun_when_check5_replace}{check5にひっかかったとき、候補となるパラメータの中からどのパラメータを代表としてとってくるか。median, max, minなどの関数を入れる。デフォルト（今までの実装）はmedian。保守的な結果を出すならmaxのほうがよいかも（エクセルではmaxに近い値が出力されるみたい）}
}
\description{
再生産関係推定が収束しているかや最適解を得られているかを診断する関数
}
