% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_tools.R
\name{check_ar_params}
\alias{check_ar_params}
\title{Check args for auto-regression estimation}
\usage{
check_ar_params(method, ar)
}
\arguments{
\item{method}{Character either 'L1' or 'L2'}

\item{ar}{Character either 'inside' or 'outside'. NULL accepted}
}
\description{
Check args for auto-regression estimation
}
