% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\name{check_consistent_w}
\alias{check_consistent_w}
\title{fit.SRとfit.SRregimeの重み付け（どのデータを使うか）の設定方法は、引数wで与える場合とSRdata$weightで与える場合の２パターンある。両者が矛盾なく設定されているかを確認するための関数}
\usage{
check_consistent_w(w, SRdata)
}
\description{
fit.SRとfit.SRregimeの重み付け（どのデータを使うか）の設定方法は、引数wで与える場合とSRdata$weightで与える場合の２パターンある。両者が矛盾なく設定されているかを確認するための関数
}
