% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{check_MSE_sd0}
\alias{check_MSE_sd0}
\title{MSEの計算が正しいかを確認するための関数}
\usage{
check_MSE_sd0(
  data_future,
  data_MSE = NULL,
  nsim_for_check = 10000,
  tol = c(0.01, 0.01, 0.01)
)
}
\description{
1. sd=0の場合の結果の比較
2. MSE_nsim=2にしてMSE_sd=0にしても良いかどうか
}
