% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_tools.R
\name{check_sr_params}
\alias{check_sr_params}
\title{Check args for auto-regression estimation}
\usage{
check_sr_params(type, method)
}
\arguments{
\item{type}{Character one of 'HS', 'RI' or 'BH'}

\item{method}{Character either 'L1' or 'L2'}
}
\description{
Check args for auto-regression estimation
}
