% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{colorize_table}
\alias{colorize_table}
\title{値の大小に応じて表の背景にグラデーションをつける}
\usage{
colorize_table(beta_table, color)
}
\arguments{
\item{beta_table}{\code{format_beta_table}で整形したβの表}

\item{color}{表の背景となる任意の色}
}
\description{
値の大小に応じて表の背景にグラデーションをつける
}
