% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{compare_future_performance}
\alias{compare_future_performance}
\title{短期的将来予測における複数の管理方策のパフォーマンスを比較する表を出力する}
\usage{
compare_future_performance(
  future_list,
  res_vpa,
  res_MSY,
  biomass.unit = 1000,
  is_MSE = FALSE,
  ...
)
}
\arguments{
\item{future_list}{将来予測の結果のリスト}

\item{res_vpa}{VPAの結果}

\item{...}{get_performanceで必要な引数}
}
\description{
短期的将来予測における複数の管理方策のパフォーマンスを比較する表を出力する
}
