% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{compare_kobeII}
\alias{compare_kobeII}
\title{複数のKobe II tableの結果を重ね書きする}
\usage{
compare_kobeII(
  kobeII_list,
  target_stat = c("prob.over.ssbtarget", "prob.over.ssblimit", "prob.over.ssbban"),
  legend_position = "top",
  target_beta = 0.8
)
}
\arguments{
\item{kobeII_list}{betaをいろいろ変えたbeta.simulationの結果のオブジェクトのリスト}

\item{target_stat}{目的とする統計量。デフォルトは"prob.over.ssbtarget"と"prob.over.ssblimit"。beta.simulationの返り値オブジェクトのリストの名前を必要に応じてとってくる}

\item{target_beta}{kobeII table から取り出すベータの値。デフォルトは0.8。}

\item{legend.position}{凡例の位置}
}
\description{
複数のKobe II tableの結果を重ね書きする
}
\examples{
\dontrun{
}

}
