% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{compare_MSY}
\alias{compare_MSY}
\title{複数のMSYの推定結果を重ね書きする}
\usage{
compare_MSY(MSYlist, legend.position = "top", MSYname = NULL,
  ncol = 2)
}
\arguments{
\item{MSYlist}{est.MSYの返り値をリストにしたもの; 単独でも可}

\item{MSYname}{凡例につけるMSYのケースの名前。MSYlistと同じ長さにしないとエラーとなる}
}
\description{
複数のMSYの推定結果を重ね書きする
}
\examples{
\dontrun{
data(res_MSY)
compare_MSY(list(res_MSY, res_MSY))
}

}
