% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{compare_SRfit}
\alias{compare_SRfit}
\title{複数の再生産関係を比較する関数}
\usage{
compare_SRfit(SRlist, biomass.unit = 1000, number.unit = 1000)
}
\arguments{
\item{SRlist}{再生産関係の推定結果のリスト。}

\item{biomass.unit}{資源量の単位}

\item{number.unit}{尾数の単位}
}
\description{
複数の再生産関係を比較する関数
}
\examples{
\dontrun{
data(res_sr_HSL1)
data(res_sr_HSL2)

(g1 <- compare_SRfit(list(HSL1=res_sr_HSL1, HSL2=res_sr_HSL2),
                     biomass.unit=1000, number.unit=1000))

}

}
