% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{convert_future_list_table}
\alias{convert_future_list_table}
\title{future_vpaの結果オブジェクトのリストをtibble形式に変換する関数}
\usage{
convert_future_list_table(fout_list, name_vector = NULL, label = "tmp")
}
\arguments{
\item{fout_list}{future_vpaの結果のオブジェクトのリスト}
}
\description{
make_kobeII_tableに入れるオブジェクトを作る
}
