% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{convert_future_table}
\alias{convert_future_table}
\title{future.vpaの結果オブジェクトをtibble形式に変換する関数}
\usage{
convert_future_table(fout, label = "tmp")
}
\arguments{
\item{fout}{future.vpaの結果のオブジェクト}
}
\description{
future.vpaの結果オブジェクトをtibble形式に変換する関数
}
