% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{convert_hvpa_tibble}
\alias{convert_hvpa_tibble}
\title{半期VPAの結果オブジェクトをtibble形式に変換する関数}
\usage{
convert_hvpa_tibble(vpares, SPRtarget = NULL)
}
\arguments{
\item{vpares}{vpaの結果のオブジェクト}
}
\description{
半期VPAの結果オブジェクトをtibble形式に変換する関数
}
