% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{convert_SR_tibble}
\alias{convert_SR_tibble}
\title{fit.SRの結果をtibble形式に治す}
\usage{
convert_SR_tibble(res_SR)
}
\arguments{
\item{SR_result}{fit.SRの結果のオブジェクト}
}
\description{
fit.SRの結果をtibble形式に治す
}
