% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{convert_vpa_tibble}
\alias{convert_vpa_tibble}
\title{VPAの結果オブジェクトをtibble形式に変換する関数}
\usage{
convert_vpa_tibble(vpares)
}
\arguments{
\item{vpares}{vpaの結果のオブジェクト}
}
\description{
VPAの結果オブジェクトをtibble形式に変換する関数
}
