% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{corSR}
\alias{corSR}
\title{再生産関係の推定パラメータの相関を出力する関数}
\usage{
corSR(resSR)
}
\arguments{
\item{resSR}{\code{fit.SR}または\code{fit.SRregime}のオブジェクト}
}
\value{
以下の要素からなるリスト
\describe{
\item{\code{hessian}}{ヘッセ行列}
\item{\code{cov}}{推定されたパラメータの分散共分散行列}
\item{\code{cor}}{推定されたパラメータの相関行列}
}
}
\description{
再生産関係の推定パラメータの相関を出力する関数
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
                out.AR = TRUE)
corRes = corSR(resSR)
corRes$cor
}
}
