% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{data.handler.hay}
\alias{data.handler.hay}
\title{csvデータを読み込んでvpa.hay用のデータを作成する関数}
\usage{
data.handler.hay(
  caa1,
  caa2,
  waa1,
  waa2,
  maa,
  maa_ssb = NULL,
  waa_ssb = NULL,
  M = 0.295,
  omega1 = NULL,
  omega2 = NULL,
  index = NULL
)
}
\description{
csvデータを読み込んでvpa.hay用のデータを作成する関数
}
\details{
`data.handler`の`vpa.hay`関数バージョン。
}
