% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{data.handler}
\alias{data.handler}
\title{csvデータを読み込んでvpa用のデータを作成する関数}
\usage{
data.handler(caa, waa, maa, index = NULL, M = 0.4, maa.tune = NULL,
  waa.catch = NULL, catch.prop = NULL)
}
\arguments{
\item{caa}{catch at age}

\item{waa}{weight at age}

\item{maa}{maturity at age}
}
\description{
csvデータを読み込んでvpa用のデータを作成する関数
}
