% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{data_future_HSL1}
\alias{data_future_HSL1}
\title{data_future_HSL1}
\format{
output object from make_future_data function

output object from future_vpa
}
\usage{
data_future_HSL1

data_future_HSL1
}
\description{
example data for conducting future_vpa with res_vpa_example & res_sr_HSL1

example output of future_vpa with data_future_HSL1
}
\keyword{datasets}
