% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{derive_future_summary}
\alias{derive_future_summary}
\title{future_vpaの返り値に要約統計量を加えるための内部関数}
\usage{
derive_future_summary(res_future, target = NULL)
}
\arguments{
\item{res_future}{future_futureの返り値}

\item{target}{指定するとtargetで指定した列の値のみが抽出される。NULLの場合は全シミュレーションの平均値}
}
\description{
future_vpaの返り値に要約統計量を加えるための内部関数
}
