% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{derive_RP_value}
\alias{derive_RP_value}
\title{管理基準値表から目的の管理基準値を取り出す関数}
\usage{
derive_RP_value(refs_base, RP_name)
}
\arguments{
\item{refs_base}{est.MSYから得られる管理基準値の表}

\item{RP_name}{取り出したい管理基準値の名前}
}
\description{
管理基準値表から目的の管理基準値を取り出す関数
}
