% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_caaboot_vpa}
\alias{do_caaboot_vpa}
\title{年齢別漁獲尾数の不確実性評価の関数}
\usage{
do_caaboot_vpa(res, B_ite = 1000, B_cv = 0.2, ci_range = 0.95)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{B_ite}{ブートストラップ計算の数。デフォルトで1000。}

\item{B_cv}{乱数生成の変動係数。デフォルトは0.2。}

\item{ci_range}{信頼区間の幅。デフォルトでは0.95（95％信頼区間）}
}
\value{
返ってくる値:
    \code{plot} 親魚重量、資源尾数、資源重量それぞれについて信頼区間のプロットが得られる。
    \code{caa_boot_sample} 年齢別漁獲尾数のブートストラップ標本が得られる。
}
\description{
年齢別漁獲尾数を乱数で再生成した後、VPA計算を行う関数
}
\seealso{
https://ichimomo.github.io/frasyr/doc/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
