% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_estcheck_vpa}
\alias{do_estcheck_vpa}
\title{vpa結果の収束診断のための関数}
\usage{
do_estcheck_vpa(res, n_ite = 20, sd_jitter = 1, what_plot = NULL, TMB = FALSE)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{n_ite}{繰り返し試行数。}

\item{sd_jitter}{乱数の分散（logスケール）}

\item{what_plot}{作図したい年齢を与える。\code{term.F = "all"}出ない限り、最高齢（+グループ）の初期値の結果のみを図示する。NULLで全てのterm.fを図示、ある特定の年齢のterm.Fを図示したい場合は\code{numeric}で与える必要がある。}

\item{TMB}{繰り返し計算をTMBで行う}
}
\value{
返ってくる値:
    \code{initial_value} 乱数で生成された初期値が得られる。
    \code{p_name} 与えた初期値（最終年の記載年のF）の名前が確認できる。
    \code{value} 各試行の推定値、尤度、収束診断、ヘッセ行列が得られる。
    \code{graph} 推定値と尤度について、それぞれ図が得られる。
}
\description{
この関数ではJitter analysisを行います。
乱数で生成した初期値を使ってvpa計算を再度行い、推定結果が同一になるかを確認することで、収束の良さを測る診断法です。
}
\seealso{
https://ichimomo.github.io/frasyr/doc/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
