% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_jackknife_vpa}
\alias{do_jackknife_vpa}
\title{ジャックナイフ法を行う関数}
\usage{
do_jackknife_vpa(res)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}
}
\value{
返ってくる値:
    \code{JKplot_abund} ジャックナイフ法を行った資源尾数の時系列プロット。
    \code{JKplot_ssb} ジャックナイフ法を行った親魚重量の時系列プロット。
    \code{JKplot_biomass} ジャックナイフ法を行った資源重量の時系列プロット。
    \code{JKplot_par} ジャックナイフ法を行ったターミナルFの時系列プロット。
    \code{data1} 資源尾数の時系列プロットに用いたtidyデータ。
    \code{data2} ターミナルFのプロットに用いたtidyデータ。
}
\description{
モデル診断法の一つです。影響力の強いデータや外れ値の検出に用いることができます。
}
\author{
濵邉昂平, 市野川桃子
}
