% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_retrospective_vpa}
\alias{do_retrospective_vpa}
\title{VPAのレトロスペクティブ解析結果を自動生成する関数}
\usage{
do_retrospective_vpa(
  res,
  n_retro = 5,
  b_reest = FALSE,
  what_plot = c("SSB", "biomass", "Recruitment", "fish_number", "fishing_mortality"),
  ncol = 3
)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{n_retro}{レトロスペクティブ解析でさかのぼる年数}

\item{b_reest}{bをレトロスペクティブ解析中で再推定するか}

\item{what_plot}{作図したい項目を選べる。\code{NULL}の場合、全て（SSB, biomass, U, catch, Recruitment, fish_number, fishing_mortality, weight, maturity, catch_number）をプロットする。}

\item{ncol}{作図の列数。標準で5列なので、\code{what_plot}の数が5以下の場合は適宜変えた方がよい。}
}
\value{
返ってくる値:
    \code{result} 感度分析の結果が\code{list}型式で得られる。
    \code{mohn_rho}
    \code{graph} 図が得られる。
}
\description{
VPAのレトロスペクティブ解析結果を自動生成する関数
}
\seealso{
レトロスペクティブ解析について:  \code{\link{retro.est}}
作図について: \code{\link{plot_vpa}}
https://ichimomo.github.io/frasyr/doc/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
