% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_sensitivity_vpa}
\alias{do_sensitivity_vpa}
\title{VPAの感度分析のための関数}
\usage{
do_sensitivity_vpa(
  res,
  what_replace,
  value,
  what_plot = NULL,
  ncol = 5,
  plot_year = NULL
)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{what_replace}{感度分析の対象。"M"（自然死亡係数）, "maa"（年齢別成熟率）, "waa"（年齢別体重）, "waa.catch"（漁獲物の年齢別体重）, "alpha"（+グループA歳とA-1歳の漁獲係数の比）, "tuning"（最終年の漁獲係数の扱い）, "lambda"（リッジVPAのペナルティ）,"est.method"（推定方法）, "b"（hyperstability/depletion）から選べる。}

\item{value}{感度分析の対象の値。ベースケースの結果に一定割合をかけたい場合は\code{numeric}型（"M", "waa", "waa.catch", "alpha"で使用可）、ベースケースの結果とは異なる表を入れたい場合は\code{list}型に\code{matrix}型を（"maa", "waa", "waa.catch"で使用可）入れてください。"tuning"を選んだ場合、資源量指数を用いていないVPA結果なら、tf.yearの期間を入れること。}

\item{what_plot}{作図したい項目を選べる。引数を与えない場合（\code{NULL}）、全て（SSB, biomass, U, catch, Recruitment, fish_number, fishing_mortality, weight, maturity, catch_number）をプロットする。}

\item{ncol}{作図の列数。標準で5列なので、\code{what_plot}の数が5以下の場合は適宜変えた方がよい。}
}
\value{
返ってくる値:
    \code{result} 感度分析の結果が\code{list}型式で得られる。
    \code{graph} 図が得られる。
}
\description{
VPAの感度分析のための関数
}
\details{
"M"（自然死亡係数）:
\code{value}にはベースケースの結果にかけたい割合を\code{numeric}型で記述のこと。

"maa"（年齢別成熟率）:
\code{value}にはベースケースと同じ行列数の\code{matrix}型で記述のこと。

"waa"（年齢別体重）, "waa.catch"（漁獲物の年齢別体重）:
\code{value}にはベースケースの結果にかけたい割合を\code{numeric}型、またはベースケースと同じ行列数の\code{matrix}型で記述のこと。

"alpha"（+グループA歳とA-1歳の漁獲係数の比）:
\code{value}には仮定したい比率を\code{numeric}型で記述のこと。

"tuning"（最終年の漁獲係数の扱い）:
resの計算方法に応じて\code{value}に入れる値が変わる。
チューニングしていないVPAの場合、感度分析も同様にチューニングしないVPAを異なるターミナルFの仮定で行う。\code{value}は仮定したい\code{tf.year}を\code{list}型で記述のこと。
全F推定の場合、感度分析では選択率更新法を行う。\code{value}は仮定したい\code{tf.year}を\code{list}型で記述のこと。
選択率更新法の場合、感度分析では全F推定を行う。\code{value}は入力不要である。
リッジVPAの場合、感度分析では異なるlambdaでの解析を行う。\code{value}は仮定したい\code{tf.year}を\code{list}型で記述のこと。

"lambda"（リッジVPAのペナルティ）: \code{value}は与えたいlambdaを\code{numeric}型で記述のこと。

"est.method"（推定方法）: \code{value}は入力不要である。

"b"(hyperstability/depletion):
bを推定した場合、bを推定しない(b=1)の感度分析を行う。
bを1以外の値で固定した、あるいはbを考慮しなかった場合、\code{value="b.est"}とするとbの推定を、仮定したいbを\code{list}で与えるとbを固定して考慮出来る。
}
\seealso{
vpa計算について:  \code{\link{vpa}}
作図について: \code{\link{plot_vpa}}
https://ichimomo.github.io/frasyr/doc/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
