% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{est.MSY}
\alias{est.MSY}
\title{VPAの結果からMSY推定を行う}
\usage{
est.MSY(vpares, farg, seed = farg$seed, eyear = 0, FUN = mean,
  N = 1000, onlylower.pgy = FALSE, optim.method = "optimize",
  calc.yieldcurve = TRUE, trace.multi = c(seq(from = 0, to = 0.9, by =
  0.1), 1, seq(from = 1.1, to = 2, by = 0.1), 3:5, 7, 20, 100),
  is.plot = TRUE, PGY = NULL, B0percent = NULL, Bempirical = NULL,
  nyear = NULL)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト}

\item{frag}{MSY計算時の将来予測で用いる引数のリスト}
}
\description{
VPAの結果からMSY推定を行う
}
