% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{est_MSYRP}
\alias{est_MSYRP}
\title{future_vpaを使ってMSY管理基準値などを計算するwrapper関数}
\usage{
est_MSYRP(
  data_future,
  ncore = 0,
  optim_method = "R",
  compile_tmb = FALSE,
  candidate_PGY = c(0.1, 0.6),
  only_lowerPGY = "lower",
  candidate_B0 = -1,
  candidate_Babs = -1,
  candidate_Fbase = -1,
  calc_yieldcurve = TRUE,
  trace.multi = unique(sort(c(0.001, seq(from = 0, to = 2, by = 0.1), 10, 100))),
  select_Btarget = 0,
  select_Blimit = 0,
  select_Bban = 0,
  multi_upper_PGY = 10,
  obj_stat = "mean"
)
}
\arguments{
\item{data_future}{make_future_dataの返り値}

\item{candidate_PGY}{PGYの計算候補}

\item{candidate_B0}{b0の計算候補}

\item{candidate_Babs}{Babsの計算候補}

\item{trace.multi}{（このベクトル）×（管理基準値として計算されるF）の平行状態の資源状態などを計算する。HCR_catchのプロットするときに、この値をもっと細かく設定することが必要になってくるかも}

\item{multi_upper_PGY}{PGYを計算するときの探索範囲の上限。デフォルトは10だが、うまくいかない場合にはこの数字を少し変えてみるとよいかも。}
}
\description{
future_vpaを使ってMSY管理基準値などを計算するwrapper関数
}
