% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{export_formattable}
\alias{export_formattable}
\title{表を画像として保存}
\usage{
export_formattable(table, file, width = "100\%", height = NULL,
  background = "white", delay = 0.1)
}
\arguments{
\item{table}{ファイルとして保存したい表}
}
\description{
表を画像として保存
}
\examples{
\dontrun{
your_table \%>\%
 export_formattable(file = "foo.png")
}
}
