% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{export_kobeII_table}
\alias{export_kobeII_table}
\title{kobeIItableから任意の表を取得し画像として保存}
\usage{
export_kobeII_table(name, divide_by, color, fname, kobeII_table)
}
\description{
kobeIItableから任意の表を取得し画像として保存
}
