% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{export_kobeII_tables}
\alias{export_kobeII_tables}
\title{β調整による管理効果を比較する表を画像として一括保存}
\usage{
export_kobeII_tables(kobeII_table, fname_ssb = "tbl_ssb.png",
  fname_catch = "tbl_catch.png",
  fname_ssb_above_target = "tbl_ssb>target.png",
  fname_ssb_above_limit = "tbl_ssb>limit.png")
}
\arguments{
\item{fname_ssb}{「平均親魚量」の保存先ファイル名}

\item{fname_catch}{「平均漁獲量」の保存先ファイル名}

\item{fname_ssb_above_target}{「親魚量が目標管理基準値を上回る確率」の保存先ファイル名}

\item{fname_ssb_above_limit}{「親魚量が限界管理基準値を上回る確率」の保存先ファイル名}
}
\description{
β調整による管理効果を比較する表を画像として一括保存
}
\examples{
\dontrun{
export_kobeII_tables(kobeII.table)
}
}
