% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{export_tables}
\alias{export_tables}
\title{Export tables to csv}
\usage{
export_tables(to, ...)
}
\arguments{
\item{to}{Name of file}

\item{...}{Table objects

\dontrun{
export_tables(to = "hoge.csv",
              table1, table2, table3)
}}
}
\description{
Export tables to csv
}
