% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_tools.R
\name{extract_xaa}
\alias{extract_xaa}
\title{Extract value at age from vpadata}
\usage{
extract_xaa(vpadata, x, year, mean_by = NULL)
}
\arguments{
\item{vpadata}{Object return from \code(vpa)}

\item{x}{Character one of 'f', 'c', 'w', 'n', 's' or 'wc'}

\item{mean_by}{If non-null, return mean value:
\describe{
  \item{"age"}{Return values averaged by age}
  \item{"year"}{Return values averaged by year}
}}
}
\description{
Extract value at age from vpadata
}
