% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\name{fit.SR}
\alias{fit.SR}
\title{加入の残差の自己相関を考慮した再生産関係の推定
L1ノルム（最小絶対値）も推定できる (sigmaはSD)
TMB = TRUEでmarginal likelihood (.cppファイルが必要)}
\usage{
fit.SR(SRdata, SR = "HS", method = "L2", AR = 1, TMB = FALSE,
  hessian = FALSE, w = rep(1, length(SRdata$year)), length = 20,
  max.ssb.pred = 1.3, p0 = NULL, out.AR = FALSE)
}
\arguments{
\item{SRdata}{get.SRdataで作成した再生産関係データ}
}
\description{
加入の残差の自己相関を考慮した再生産関係の推定
L1ノルム（最小絶対値）も推定できる (sigmaはSD)
TMB = TRUEでmarginal likelihood (.cppファイルが必要)
}
