% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{fit.SRalpha}
\alias{fit.SRalpha}
\title{再生産関係の推定}
\usage{
fit.SRalpha(
  SRdata,
  SR = "HS",
  alpha = 0,
  hessian = FALSE,
  w = rep(1, length(SRdata$R)),
  length = 20,
  max.ssb.pred = 1.3,
  p0 = NULL,
  rep.opt = TRUE
)
}
\arguments{
\item{SRdata}{\code{get.SRdata}で作成した再生産関係データ}

\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker)}

\item{alpha}{alpha:(1-alpha)の比でL1:L2を混ぜる}

\item{length}{初期値を決める際のgridの長さ}

\item{p0}{\code{optim}で設定する初期値}

\item{rep.opt}{\code{TRUE}で\code{optim}による最適化を収束するまで繰り返す}
}
\description{
3種類の再生産関係の推定を、最小二乗法か最小絶対値法で、さらに加入の残差の自己相関を考慮して行うことができる
}
