% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-patchSR.R
\encoding{UTF-8}
\name{fit.SRmig}
\alias{fit.SRmig}
\title{ある系群から別の系群への移入が想定される場合の再生産関係の推定（マアジ用）}
\usage{
fit.SRmig(
  SRdata1,
  SRdata2,
  SR = c("HS", "HS"),
  AR = c(FALSE, FALSE),
  mig_year = NULL,
  mig_type = 0,
  one_patch = FALSE,
  sigma_omega_ratio = 0,
  omega_fix = NA,
  phase = FALSE,
  p0_list = NULL
)
}
\arguments{
\item{SRdata1}{移出が生じている系群の\code{SRdata}のオブジェクト}

\item{SRdata2}{移入が生じている系群の\code{SRdata}のオブジェクト}

\item{mig_year}{移出入が生じている期間}

\item{mig_type}{0: 移出入なし, 1: 一定の割合で移入が生じる, 2: 平均的な割合からランダムに変動する, 3:移出する割合が自己回帰モデルで変化する, 4: 移出する割合がランダムウォーク}
}
\description{
ある系群から別の系群への移入が想定される場合の再生産関係の推定（マアジ用）
}
