% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{fit.SRregime}
\alias{fit.SRregime}
\title{レジーム分けを考慮した再生産関係の推定}
\usage{
fit.SRregime(
  SRdata,
  SR = "HS",
  method = "L2",
  regime.year = NULL,
  regime.key = 0:length(regime.year),
  regime.par = c("a", "b", "sd"),
  use.fit.SR = TRUE,
  length = 10,
  p0 = NULL,
  w = rep(1, length(SRdata$R)),
  max.ssb.pred = 1.3,
  hessian = FALSE
)
}
\arguments{
\item{SRdata}{\code{get.SRdata}で作成した再生産関係データ}

\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker)}

\item{method}{最適化法（\code{"L2"}: 最小二乗法, \code{"L1"}: 最小絶対値法）}

\item{regime.year}{レジームが変わる年}

\item{regime.key}{レジームのパターンを表す(\code{0:2}だとA->B->Cで、\code{c(0,1,0)}だとA->B->Aのようなパターンとなる)}

\item{regime.par}{レジームによって変化するパラメータ(\code{c("a","b","sd")}の中から選ぶ)}

\item{length}{初期値を決める際のgridの長さ}

\item{p0}{\code{optim}で設定する初期値}

\item{sd.obs}{L1のときSDをRMSEに置き換えるか}
}
\value{
以下の要素からなるリスト
\describe{
\item{\code{input}}{使用した引数のリスト}
\item{\code{pars}}{推定されたパラメータ}
\item{\code{opt}}{\code{optim}の結果オブジェクト}
\item{\code{resid}}{再生産関係から予測値からの加入量の残差}
\item{\code{loglik}}{対数尤度}
\item{\code{k}}{推定したパラメータ数}
\item{\code{AIC}}{AIC}
\item{\code{AICc}}{AICc}
\item{\code{BIC}}{BIC}
\item{\code{regime_pars}}{レジームごとの推定パラメータ}
\item{\code{regime_resid}}{レジームごとの残差}
\item{\code{pred}}{レジームごとの各親魚量に対する加入量の予測値}
\item{\code{pred_to_obs}}{観測値に対する予測値}
\item{\code{summary_tbl}}{観測値と予測値を合わせた表}
}
}
\description{
レジームシフトが生じた年やレジームであるパラメータが共通する場合やレジームのパターンがA->B->CなのかA->B->Aなのか等が検討できる
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSRregime <- fit.SRregime(SRdata, SR="HS", method="L2",
                            regime.year=c(1995,2005), regime.key=c(0,1,0),
                            regime.par = c("a","b","sd")[2:3])
resSRregime$regime_pars
}
}
