% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{fit_draw_SRline}
\alias{fit_draw_SRline}
\title{Fit and draw SR line(s) over observed stock-recruitment plots}
\usage{
fit_draw_SRline(vpares, SR, method, show.year = FALSE)
}
\arguments{
\item{SR}{Vector of multiple SR model described in \code{\link{fit.SR}}}

\item{method}{Vector of multiple method described in \code{\link{fit.SR}}}
}
\description{
Fit and draw SR line(s) over observed stock-recruitment plots
}
\examples{
\dontrun{
draw_SRline(vpares = res_vpa,
            SR = c("HS", "RI", "BH"),
            method = c("L1", "L2"))
}
}
