% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{format_beta_table}
\alias{format_beta_table}
\title{kobeIItableから取り出した表を整形}
\usage{
format_beta_table(beta_table, divide_by = 1, round = TRUE)
}
\arguments{
\item{beta_table}{\code{pull_var_from_kobeII_table}で取得した表}

\item{divide_by}{表の値をこの値で除する．トンを千トンにする場合には1000}

\item{round}{TRUEなら値を丸める．漁獲量は現状整数表示なのでデフォルトはTRUE}
}
\description{
- 報告書に不要な列を除去する
- 単位を千トンに変換
}
