% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{generate_sr_weight}
\alias{generate_sr_weight}
\title{Generates weights for years to use SR relationship}
\usage{
generate_sr_weight(srdata, using = NULL, n_yrs_to_remove = NULL)
}
\arguments{
\item{srdata}{Object returned by \code{get.SRdata}}

\item{using}{Years to use either "allyears" or numeric years}

\item{n_yrs_to_remove}{N of recent years to omit

\dontrun{
srdata <- get.SRdata(result_vpa)
generate_sr_weight(using = "allyear", srdata)
generate_sr_weight(using = 1982:1990, srdata)
generate_sr_weight(using = 1982:1990, srdata)
generate_sr_weight(n_yrs_to_remove = 3, srdata)
}}
}
\description{
Generates weights for years to use SR relationship
}
