% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{Generation.Time}
\alias{Generation.Time}
\title{VPAの結果に格納されている生物パラメータから世代時間を計算}
\usage{
Generation.Time(
  vpares,
  maa.year = 2014:2015,
  maa = NULL,
  M.year = 2014:2015,
  M = NULL,
  Plus = 19
)
}
\description{
VPAの結果に格納されている生物パラメータから世代時間を計算
}
