% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{get_performance}
\alias{get_performance}
\title{将来予測の結果のリストを入れると、代表的なパフォーマンス指標をピックアップする}
\usage{
get_performance(
  future_list,
  res_vpa,
  ABC_year = 2021,
  is_MSE = FALSE,
  indicator_year = c(0, 5, 10),
  Btarget = 0,
  Blimit = 0,
  Bban = 0,
  type = c("long", "wide")[1],
  biomass.unit = 10000,
  ...
)
}
\arguments{
\item{future_list}{future_vpaまたはfuture.vpaの返り値のリスト}

\item{res_vpa}{vpaの返り値}

\item{ABC_year}{特にABCに注目したい年}

\item{is_MSE}{MSEの結果を使うかどうか。MSEの結果の場合には、ABCや加入尾数の真の値との誤差も出力する}

\item{indicator_year}{ABC_yearから相対的に何年後を指標として取り出すか}

\item{Btarget}{目標管理基準値の値}

\item{Blimit}{限界管理基準値の値}

\item{Bban}{禁漁水準の値}

\item{type}{出力の形式。"long"は縦長（ggplotに渡すときに便利）、"wide"は横長（数字を直接比較するときに便利）}

\item{biomass.unit}{資源量の単位}
}
\description{
将来予測の結果のリストを入れると、代表的なパフォーマンス指標をピックアップする
}
