% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{get_RP}
\alias{get_RP}
\title{管理基準値表から目的の管理基準値を取り出す関数(簡易版)}
\usage{
get_RP(res_MSY, type, name = TRUE, def = TRUE)
}
\arguments{
\item{res_MSY}{est.MSYの帰り値}

\item{type}{res_MSY$summaryの列名（Fの場合だけ、"F"とする）}

\item{name}{RP.definitionで取り出す場合のラベル。使わない場合はTRUE}
}
\description{
管理基準値表から目的の管理基準値を取り出す関数(簡易版)
}
