% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_tools.R
\name{give_agename}
\alias{give_agename}
\title{Give age names starts from zero and ends with plus group}
\usage{
give_agename(x)
}
\arguments{
\item{x}{object to name
give_agename(1:3)
give_agename(-1:2)
give_agename(c(1.234, 2.345))}
}
\description{
Give age names starts from zero and ends with plus group
}
