% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{hmm_SR}
\alias{hmm_SR}
\title{再生産関係のレジームシフトを隠れマルコフモデルで推定する}
\usage{
hmm_SR(
  SRdata,
  SR = "BH",
  k_regime = 2,
  gamma = 0.01,
  b_range = NULL,
  p0 = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{SRdata}{\code{get.SRdata}で作成した再生産関係データ}

\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker, \code{"Mesnil"}: Continuous HS)}

\item{gamma}{\code{SR="Mesnil"}のときに使用するsmoothing parameter}

\item{p0}{\code{optim}で設定する初期値}
}
\description{
Tang et al. (2021) ICESJMSの論文を改変
}
