% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{hretro_est}
\alias{hretro_est}
\title{半期VPAのレトロスペクティブ解析用の関数}
\usage{
hretro_est(res, n = 5, b_fix = TRUE, rho_fix = TRUE)
}
\description{
半期VPAのレトロスペクティブ解析用の関数
}
\details{
`do_retrospective_vpa`関数内でこの関数を動かしてレトロスペクティブ解析の図示もしている。
}
