% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{hretro_plot}
\alias{hretro_plot}
\title{半期VPA専用のレトロスペクティブ解析の作図関数}
\usage{
hretro_plot(
  res,
  retro,
  start_year = 2014,
  target = "naa",
  age = 0:4,
  season = 1,
  add_mohn = TRUE,
  digits = 3,
  hanki = TRUE,
  out = "plot",
  title_size = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  size = 1.5
)
}
\description{
半期VPA専用のレトロスペクティブ解析の作図関数
}
