% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{HS.recAR}
\alias{HS.recAR}
\title{HSを仮定したときの加入関数}
\usage{
HS.recAR(ssb, vpares, rec.resample = NULL, rec.arg = list(a = 1000, b =
  1000, sd = 0.1, rho = 0, resid = 0))
}
\arguments{
\item{ssb}{親魚資源量}

\item{vpares}{VPAの出力結果}
}
\description{
HSを仮定したときの加入関数
}
