% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{hvpa_plot}
\alias{hvpa_plot}
\title{半期VPAの作図関数}
\usage{
hvpa_plot(
  res,
  res0 = NULL,
  years = NULL,
  age = NULL,
  rel_heights = c(0.7, 0.3),
  labels = c("Abundance", "Fishing Rate")
)
}
\description{
半期VPAの作図関数
}
