% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{is_autocorrelated}
\alias{is_autocorrelated}
\title{自己相関の検定}
\usage{
is_autocorrelated(x, ci = 0.95)
}
\arguments{
\item{x}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}

\item{ci}{自己相関のための信頼区間．デフォルト値は0.95}
}
\description{
自己相関の検定
}
\examples{
\dontrun{
res_vpa \%>\%
  get.SRdata() \%>\%
  fit.SR("HS", "L1") \%>\%
  pull_df_from_list("resid2") \%>\%
  is_autocorrelated()
}
}
