% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{is_resid_normdist}
\alias{is_resid_normdist}
\title{再生産関係のフィット残差の正規性検定}
\usage{
is_resid_normdist(srdata, p_threshold = 0.05)
}
\arguments{
\item{srdata}{fit.SR()の出力}

\item{p_threshold}{有意性判定の基準とするp値}
}
\description{
再生産関係のフィット残差の正規性検定
}
\examples{
\dontrun{
vpares \%>\%
  get.SRdata() \%>\%
  fit.SR(SR = "HS", method = "L1") \%>\%
  is_resid_normdist()
}`
}
