% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{jackknife.SR}
\alias{jackknife.SR}
\title{再生産関係のジャックナイフ解析}
\usage{
jackknife.SR(
  resSR,
  is.plot = TRUE,
  use.p0 = TRUE,
  output = FALSE,
  filename = "jackknife",
  ylim.range = c(0, 1.5),
  pch = 19,
  cex = 1.1,
  ...
)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{is.plot}{プロットするかどうか}

\item{use.p0}{初期値に\code{resSR}の結果を使うか否か}

\item{output}{pngファイルに出力するか否か}

\item{filename}{ファイル名}
}
\description{
結果のプロットもこの関数で行う
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
                out.AR = TRUE)
res_jackSR <- jackknife.SR(resSR,output = T)

# if calculate steepness
bio_par <- derive_biopar(res_obj=res_vpa,derive_year = 2010)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
                out.AR = TRUE,bio_par=bio_par)
res_jackSR <- jackknife.SR(resSR,output = T)
}
}
