% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.r
\name{make_future_data_boot}
\alias{make_future_data_boot}
\title{create future projection data based on bootstrap results}
\usage{
make_future_data_boot(res_vpa, vpa_boot, SR_boot, seed = 1, ...)
}
\description{
create future projection data based on bootstrap results
}
