% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{make_kobe_ratio}
\alias{make_kobe_ratio}
\title{Make Kobe ratio (?) for frasyr_tool}
\usage{
make_kobe_ratio(result_vpa, result_msy)
}
\arguments{
\item{result_vpa}{object createb dy vpa()}

\item{result_msy}{object created by script 1do_MSYest.R of SC meeting}
}
\value{
A tibble object
}
\description{
Make Kobe ratio (?) for frasyr_tool
}
