% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{make_kobeII_table_med}
\alias{make_kobeII_table_med}
\title{beta.simluationの結果などを読んで、kobeII tableに整形する関数で平均値の代わりに中央値を計算する関数}
\usage{
make_kobeII_table_med(
  kobeII_data,
  res_vpa,
  year.catch = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssb = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbtarget = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssblimit = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbban = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbmin = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  B.vector = c(Btarget = 0, Blimit = 0, Bban = 0)
)
}
\arguments{
\item{kobeII_data}{beta.simulationまたはconvert_future_list_tableの返り値}

\item{res_vpa}{VPAの結果}

\item{year.catch}{平均漁獲量の表を出力する期間。その他year.ssb(平均親魚量), year.ssbtarget(SBtargetを上回る確率)、、なども同様}

\item{B.vector}{上回っている確率を算出する親魚量のベクトル}
}
\description{
beta.simluationの結果などを読んで、kobeII tableに整形する関数で平均値の代わりに中央値を計算する関数
}
\details{
tidy形式になっているkobeII_dataにおいて、HCR_name, betaの列のラベルの組み合わせを一つの管理方式として、その管理方式ごとに少尉予測の結果を集計する
}
