% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_msytable}
\alias{make_msytable}
\title{Make summary table of MSY estimation}
\usage{
make_msytable(result_msy)
}
\arguments{
\item{result_msy}{Object of msy result}
}
\description{
Make summary table of MSY estimation
}
