% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_stock_table}
\alias{make_stock_table}
\title{Make table for stock assessment result}
\usage{
make_stock_table(result_vpa, result_msy, yr_future_start, unit = "百トン")
}
\arguments{
\item{result_vpa}{Obejct returned from \code{vpa()}}

\item{result_msy}{Object MSY result created previous SC meeting...?}

\item{yr_future_start}{Year when future projection starts}
}
\description{
Make table for stock assessment result
}
