% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_stock_table2}
\alias{make_stock_table2}
\title{Make table for stock assessment result 2}
\usage{
make_stock_table2(
  result_future0.8,
  yr_future_ABC,
  Fcurrent_per_Fmsy,
  Fcurrent_year_range
)
}
\arguments{
\item{result_future0.8}{Object returnd from adopted future projection}

\item{yr_future_ABC}{Year when ABC is calculated}

\item{Fcurrent_per_Fmsy}{Fcurrent/Fmsy}

\item{Fcurrent_year_range}{}
}
\description{
Make table for stock assessment result 2
}
