% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{out.SR}
\alias{out.SR}
\title{再生産関係の推定結果をtxtファイルに出力する関数}
\usage{
out.SR(resSR, filename = "resSR")
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{filename}{ファイル名('.txt')がつく}
}
\description{
再生産関係の推定結果をtxtファイルに出力する関数
}
