% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\name{out.vpa}
\alias{out.vpa}
\title{VPA結果をcsvファイルに出力する}
\usage{
out.vpa(res = NULL, rres = NULL, fres = NULL, ABC = NULL,
  filename = "vpa")
}
\arguments{
\item{res}{VPAの結果}

\item{rres}{ref.Fの結果}

\item{fres}{future.vpaの結果}
}
\description{
VPA結果をcsvファイルに出力する
}
