% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{pass_shapiro_and_ks}
\alias{pass_shapiro_and_ks}
\title{ベクトルxの正規性を検定する}
\usage{
pass_shapiro_and_ks(x, sd, p_threshold)
}
\arguments{
\item{x}{a numeric vector of data values. Missing values are allowed,
    but the number of non-missing values must be between 3 and 5000.}

\item{sd}{SD of pnorm function to compare}

\item{p_threshold}{有意性判定の基準とするp値}
}
\description{
Shapiro testとKolmogorov-Smirnov testの両方において
帰無仮説を棄却できなかった（母集団が正規分布と異なるとは言えない）場合に
\code{TRUE}が返る
}
