% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_influence.R
\name{plot_data_influence}
\alias{plot_data_influence}
\title{推定資源量の 1 セルが、どのデータセルにどれだけ依存しているかのヒートマップ}
\usage{
plot_data_influence(infl, n_age = 0, n_year = "last")
}
\arguments{
\item{infl}{[data_influence()] の返り値}

\item{n_age}{対象の年齢 (0 始まり)}

\item{n_year}{対象の年 (列名)。`"last"` なら終端年}
}
\value{
ggplot オブジェクト (塗り = 弾力性 d log N / d log データ)
}
\description{
推定資源量の 1 セルが、どのデータセルにどれだけ依存しているかのヒートマップ
}
