% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{plot_Fcurrent}
\alias{plot_Fcurrent}
\title{F currentをプロットする}
\usage{
plot_Fcurrent(vpares, year.range = NULL)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト}
}
\description{
F currentをプロットする
}
