% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{plot_Fref}
\alias{plot_Fref}
\title{ref.Fの出力をプロットするための関数}
\usage{
plot_Fref(rres, xlabel = "max", vline.text = c("F0.1", "Fmax",
  "Fcurrent", "Fmed"))
}
\arguments{
\item{rres}{ref.Fの出力結果}
}
\description{
ref.Fの出力をプロットするための関数
}
