% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_future_simple}
\alias{plot_future_simple}
\title{一つの将来予測の結果をプロットする（ggplotは使わず）}
\usage{
plot_future_simple(
  fres0,
  ylim.tmp = NULL,
  xlim.tmp = NULL,
  vpares = NULL,
  what = c(TRUE, TRUE, TRUE),
  conf = 0.1,
  N.line = 0,
  det.run = TRUE,
  label = c("Biomass", "SSB", "Catch"),
  is.legend = TRUE,
  add = FALSE,
  col = NULL,
  ...
)
}
\arguments{
\item{fres0}{future.vpaからの出力結果}
}
\description{
一つの将来予測の結果をプロットする（ggplotは使わず）
}
