% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_futures}
\alias{plot_futures}
\title{将来予測の複数の結果をggplotで重ね書きする}
\usage{
plot_futures(
  vpares = NULL,
  future.list = NULL,
  future.name = names(future.list),
  future_tibble = NULL,
  CI_range = c(0.1, 0.9),
  maxyear = NULL,
  minyear = NULL,
  is.plot.CIrange = TRUE,
  is.plot.CIline = TRUE,
  what.plot = c("Recruitment", "SSB", "biomass", "cbiomass", "catch", "beta_gamma", "U",
    "Fratio"),
  biomass.unit = 1,
  number.unit = 1,
  number.name = "",
  RP_name = c("Btarget", "Blimit", "Bban", "Fmsy", "MSY", "U_MSY"),
  Btarget = NA,
  Blimit = NA,
  Bban = NA,
  MSY = NA,
  Umsy = NA,
  SPRtarget = NULL,
  exclude.japanese.font = FALSE,
  average_lwd = 1,
  n_example = 3,
  example_width = 0.7,
  future.replicate = NULL,
  seed = 1,
  legend.position = "top",
  type = "detail",
  font.size = 16,
  ncol = 3,
  ncol_legend = 2,
  remove.last.vpa.year = FALSE,
  use_median = FALSE
)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト(NULLでもOK)}

\item{future.list}{将来予測の結果をリストで並べたもの(こちらは必須)}

\item{future.name}{将来予測のリストの名前。ない場合はfuture.listについている名前を使う}

\item{CI_range}{予測区間の範囲。デフォルトは８０\%でc(0.1,0.9)}

\item{maxyear}{表示する年の最大}

\item{minyear}{表示する年の最小}

\item{is.plot.CIrange}{予測区間を表示するかどうか}

\item{is.plot.CIline}{予測区間の上限と下限の線を引くか}

\item{what.plot}{Recruitment,SSB,biomass,catch,beta_gamma,U,Fratioのうち何をプロットするか。これらの文字列のベクトルで指定する}

\item{biomass.unit}{量の単位}

\item{number.unit}{尾数（加入尾数とか）のときの単位}

\item{number.name}{尾数の凡例をどのように表示するか（「億尾」とか）}

\item{RP_name}{管理基準値をどのように名前つけるか（デフォルトはc("Btarget","Blimit","Bban","Fmsy","MSY","U_MSY"))}

\item{Btarget}{Btargetの値、設定しない場合はNA（デフォルトはNA）}

\item{Blimit}{Blimitの値、設定しない場合はNA（デフォルトはNA）}

\item{Bban}{Bbanの値、設定しない場合はNA（デフォルトはNA）}

\item{MSY}{MSYの値、設定しない場合はNA（デフォルトはNA）}

\item{Umsy}{Umsyの値、設定しない場合はNA（デフォルトはNA）}

\item{SPRtarget}{MSYのときのSPRの値}

\item{exclude.japanese.font}{日本語を図中に表示しない}

\item{average_lwd}{将来予測の平均値の線の太さ. 基本は1.}

\item{n_example}{個々のシミュレーションの例を示す数}

\item{future.replicate}{どのreplicateを選ぶかを選択する。この場合n_exampleによる指定は無効になる}

\item{legend.position}{凡例の位置。top, right, left, bottomなど}

\item{font.size}{フォントの大きさ}

\item{ncol}{図を並べるときの列数}

\item{remove.last.vpa.year}{VPAの最終年のデータのプロットを除くかどうか（last.catch.zero用オプション）}

\item{use_median}{平均値の代わりに中央値をプロットする}

\item{width_example}{個々のシミュレーションをプロットする場合の線の太さ (default=0.7)}
}
\description{
将来予測の複数の結果をggplotで重ね書きする
}
