% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_futures_simple}
\alias{plot_futures_simple}
\title{Plot multiple future projection results without ggplot}
\usage{
plot_futures_simple(
  fres.list,
  conf = c(0.1, 0.5, 0.9),
  target = "SSB",
  legend.text = "",
  xlim.tmp = NULL,
  y.scale = 1,
  det.run = TRUE
)
}
\arguments{
\item{fres.list}{List of future projection results from `future.vpa`.}
}
\description{
Plot multiple future projection results without ggplot
}
