% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_HCR}
\alias{plot_HCR}
\title{HCRを書く}
\usage{
plot_HCR(
  SBtarget,
  SBlim,
  SBban,
  Ftarget,
  Fcurrent = -1,
  biomass.unit = 1,
  beta = 0.8,
  col.multi2currf = "black",
  col.SBtarget = "#00533E",
  col.SBlim = "#edb918",
  col.SBban = "#C73C2E",
  col.Ftarget = "black",
  col.betaFtarget = "gray",
  is.text = TRUE,
  RP.label = c("目標管理基準値", "限界管理基準値", "禁漁水準", "Fmsy",
    "0.8Fmsy")
)
}
\arguments{
\item{SBtarget}{目標管理基準値}

\item{SBlim}{禁漁水準}

\item{Ftarget}{Ftarget}

\item{is.text}{ラベルを記入するかどうか（FALSEにすると後で自分で書き換えられる）}

\item{RP.label}{管理基準値のラベル}
}
\description{
HCRを書く
}
