% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{plot_residual_vpa}
\alias{plot_residual_vpa}
\title{残差プロットの関数}
\usage{
plot_residual_vpa(
  res,
  index_name = NULL,
  plot_smooth = TRUE,
  plot_year = FALSE,
  plot_scale = FALSE,
  resid_CI = TRUE,
  plotAR = FALSE
)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{index_name}{作図時に各指標の名前}
}
\value{
返ってくる値:
    \code{year_resid} 残差の時系列プロット
    \code{fitting_CPUE} 観測、予測CPUEのフィッティングの図
    \code{abund_CPUE} 資源量と資源量指数の当てはまりの図。
    \code{year_sd_resid} 標準化残差の時系列プロット
    \code{gg_data} 作図に用いたデータ（tidy形式）
}
\description{
残差プロットの関数
}
\seealso{
https://ichimomo.github.io/frasyr/articles/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
