% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_SR_AReffect}
\alias{plot_SR_AReffect}
\title{ARの効果を差し引いて再生産関係をプロットする}
\usage{
plot_SR_AReffect(res)
}
\arguments{
\item{res}{fit.SRから返されるオブジェクト}
}
\description{
ARの効果を差し引いて再生産関係をプロットする
}
\examples{
\dontrun{

}

}
