% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\name{plot_SRaverage}
\alias{plot_SRaverage}
\title{モデル平均の場合の再生産関係のプロット}
\usage{
plot_SRaverage(
  data_future_MSY,
  res_vpa_updated,
  SR_plot_label,
  CI = 0.9,
  last_year_color = 5
)
}
\description{
汎用化していない（２つのモデルの平均の場合だけです）
}
