% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{plot_summary_performance}
\alias{plot_summary_performance}
\title{kobeII tableの結果をさらに要約してグラフ化する}
\usage{
plot_summary_performance(
  folder_names,
  scenario_names,
  kobe_object = NULL,
  plot_year_range,
  beta_select,
  res_MSY,
  catch_3terms = list(2022:2024, 2025:2027, 2028:2030),
  ssb_3terms = list(2023:2025, 2026:2028, 2029:2031),
  risk.pickup = NULL,
  font_size = 25,
  sort_category = NULL
)
}
\description{
漁獲量・親魚量・リスクの３つに分類する
}
