% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_assess.R
\name{predict_recruitment_from_sr_pars}
\alias{predict_recruitment_from_sr_pars}
\title{Predict recruitment from stock-recruitment parameters}
\usage{
predict_recruitment_from_sr_pars(ssb, sr_type, a, b)
}
\arguments{
\item{ssb}{Numeric vector of spawning stock biomass.}

\item{sr_type}{Numeric vector or scalar. `1` = HS, `2` = BH, `3` = RI.}

\item{a, b}{Numeric vectors or scalars of stock-recruitment parameters.}
}
\value{
Numeric vector of predicted recruitment.
}
\description{
This helper evaluates stock-recruitment relationships using the parameter
coding stored in `SR_mat`. Currently, Hockey-stick (`SR_type = 1`),
Beverton-Holt (`SR_type = 2`), and Ricker (`SR_type = 3`) are supported.
}
