% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{prof.lik}
\alias{prof.lik}
\title{profile likelihood}
\usage{
prof.lik(Res, a = Res$pars$a, b = Res$pars$b, sd = Res$pars$sd,
  rho = ifelse(Res$input$out.AR, 0, Res$pars$rho))
}
\arguments{
\item{Res}{再生産関係（\code{fit.SR()}）の結果オブジェクト}
}
\description{
profile likelihood
}
