% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{prof.likSR}
\alias{prof.likSR}
\title{プロファイル尤度を計算する関数}
\usage{
prof.likSR(
  resSR,
  output = FALSE,
  filename = "Profile_Likelihood",
  a_range = NULL,
  b_range = NULL,
  HS_b_restrict = TRUE,
  length = 50
)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{output}{pngファイルに出力するか否か}

\item{filename}{ファイル名}

\item{a_range}{推定値に\code{a_range}を乗じた範囲で尤度計算を行う (\code{NULL}のとき\code{c(0.5,2)})}

\item{b_range}{推定値に\code{a_range}を乗じた範囲で尤度計算を行う (\code{NULL}のとき\code{c(0.5,2)})}

\item{HS_b_restrict}{Hockey-StickのときにbをSSBの観測値の範囲にするか否か (\code{b_range}より優先される)}

\item{length}{範囲を区切る数}
}
\description{
結果のプロットもこの関数で行う
}
