% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{profile_likelihood.vpa}
\alias{profile_likelihood.vpa}
\title{VPAの推定値についてprofile likelihood (one parameter)を実施する}
\usage{
profile_likelihood.vpa(res, Alpha = 0.95, min.p = 1e-06, max.p = 1,
  L = 20, method = "ci")
}
\arguments{
\item{res}{vpa関数の出力値}
}
\description{
VPAの推定値についてprofile likelihood (one parameter)を実施する
}
