% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{pull_df_from_list}
\alias{pull_df_from_list}
\title{Util functions for select SRmodels
Pull data frame from list}
\usage{
pull_df_from_list(list, dfname)
}
\arguments{
\item{list}{List object such as VPA result}

\item{dfname}{Name of data frame such as "baa" in VPA result}
}
\description{
Util functions for select SRmodels
Pull data frame from list
}
