% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{pull_var_from_kobeII_table}
\alias{pull_var_from_kobeII_table}
\title{kobeIItable から任意の表を指名して取り出す}
\usage{
pull_var_from_kobeII_table(kobeII_table, name)
}
\arguments{
\item{kobeII_table}{\code{make_kobeII_table}の出力}

\item{name}{\code{kobeII_table}の要素名}
}
\description{
kobeIItable から任意の表を指名して取り出す
}
