% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{rank_HCR}
\alias{rank_HCR}
\title{リスクテーブルを作る}
\usage{
rank_HCR(
  summary_HCR,
  target_threshold,
  risk_threshold,
  target_prob_name = "SBtar_prob",
  risk_prob_name = "SBlim_prob"
)
}
\arguments{
\item{target_threshold}{カテゴリ分けに使うSSB>SSBtargetの確率。1番目がbeta=0.8のときの値、2番目が50％のときの値（なのでだいたい50％になるはず）}

\item{limit_threshold}{カテゴリ分けに使うSSBall>SSBthresholdの確率。1番目がbeta=0.8のときの値、2番目が50％のときの値（なので1番目よりも2番めのほうが小さいはず）}
}
\description{
リスクテーブルを作る
}
