% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{res_sr_HSL1}
\alias{res_sr_HSL1}
\title{res_sr_HSL1}
\format{
output object from est_SR function
}
\usage{
res_sr_HSL1
}
\description{
example stock-recruitment result with res_vpa_example data (HS & L1)
}
\keyword{datasets}
