% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{res_vpa_example}
\alias{res_vpa_example}
\title{vpa_example}
\format{
output object from VPA function
}
\usage{
res_vpa_example
}
\description{
example VPA result originally from Maaji Pacific VPA result (non tuning)
}
\keyword{datasets}
