% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{res_vpa_example2}
\alias{res_vpa_example2}
\title{vpa_example2}
\format{
output object from VPA function
}
\usage{
res_vpa_example2
}
\description{
example VPA result originally from Maaji Pacific VPA result (with tuning)
the tuning data is not real data, but ones generated from lognormal error distribution (see raw-data/process.R )
}
\keyword{datasets}
