% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{residual_plot}
\alias{residual_plot}
\title{半期VPA専用の残差プロット作図の関数}
\usage{
residual_plot(
  res,
  rel_heights = c(0.6, 0.4),
  add_log_plot = TRUE,
  out = "plot",
  abund_name = NULL,
  stat = TRUE
)
}
\description{
半期VPA専用の残差プロット作図の関数
}
