% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{retrieve_input}
\alias{retrieve_input}
\title{Retrieve function argument settings to reuse}
\usage{
retrieve_input(result, new_sd = NULL)
}
\arguments{
\item{result}{objects created by make_future_data()}
}
\description{
Retrieve function argument settings to reuse
}
