% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{retro.est}
\alias{retro.est}
\title{レトロスペクティブ解析の実施}
\usage{
retro.est(
  res,
  n = 5,
  stat = "mean",
  init.est = FALSE,
  b.fix = TRUE,
  remove.maxAgeF = FALSE,
  ssb.forecast = FALSE,
  sel.mat = NULL,
  grid.add.ini = NULL,
  grid.init = NULL,
  remove.short.index = FALSE
)
}
\arguments{
\item{res}{VPAの出力}

\item{n}{除く年数}

\item{b.fix}{b推定してる場合にbを固定するかどうか}

\item{remove.maxAgeF}{Mohn's rhoを計算する際に最高齢のFを除くか（alphaを仮定して計算していることが多いから）}

\item{ssb.forecast}{Mohn's rhoを計算する際にSSBは1年後を計算するか(last.catch.zero=TRUEのときのみ有効)}

\item{grid.add.ini}{\code{add.p.ini}をgridで変えて初期値を事前に探索する}

\item{grid.init}{\code{p.init}でgridを変えて初期値を事前に探索する}

\item{remove.short.index}{年数が2年以下になった指標値を除いて計算する}
}
\description{
レトロスペクティブ解析の実施
}
