% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{RPS.simple.rec}
\alias{RPS.simple.rec}
\title{期間内のRPSをサンプリングするときの加入関数}
\usage{
RPS.simple.rec(ssb, vpares, rec.arg = list(rps.year = NULL, upper.ssb =
  Inf, upper.recruit = Inf, sample.year = NULL, bias.correction = TRUE,
  rpsmean = FALSE), deterministic = FALSE, rec.resample = NULL)
}
\description{
期間内のRPSをサンプリングするときの加入関数
}
