% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{sample_backward}
\alias{sample_backward}
\title{過去にさかのぼってブロックサンプリングをおこなう}
\usage{
sample_backward(residual, n, duration)
}
\arguments{
\item{residual}{リサンプリングする残差}

\item{n}{将来にわたって何年分のリサンプリング残差を作るか}

\item{duration}{1ブロックの年の長さ}
}
\description{
過去にさかのぼってブロックサンプリングをおこなう
}
