% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{set_f_current}
\alias{set_f_current}
\title{Set F current by direct input or referencing VPA result}
\usage{
set_f_current(manual = NULL, f_years = NULL, s_years = NULL, vpadata = NULL)
}
\arguments{
\item{manual}{Numeric vector of F values such as
c(0.123, 1.234, 2.345, 3.456)}

\item{f_years}{Years to calculate mean F values}

\item{s_years}{Years to calculate selectivity}

\item{vpadata}{VPA object created by \code{vpa()}}
}
\value{
Numeric vector
}
\description{
Set F current by direct input or referencing VPA result
}
\examples{
\dontrun{
vpadata <- vpa("some_setting")

set_f_current(f_years = 2015:2017,
              vpadata = vpadata)

set_f_current(f_years = 2015:2017,
              s_years = 2010:2016, # Use different years to calculate
              vpadata = vpadata)

set_f_current(
  f_years = -4:-2, # (N - 1) years before the latest (latest: -1)
  vpadata = vpadata),
set_f_current(f_years = -4:-2,
              f_years = -10:-1,
              vpadata = vpadata),
}
}
