% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tmb.r
\name{set_SR_mat_lognormal}
\alias{set_SR_mat_lognormal}
\title{対数正規分布の残差分布を作る関数}
\usage{
set_SR_mat_lognormal(res_vpa, res_SR, SR_mat, seed_number,
  start_random_rec_year_name, future_initial_year)
}
\description{
再生産関係は一定
}
