% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{specify.Mesnil.gamma}
\alias{specify.Mesnil.gamma}
\title{再生産関係推定でSR＝HSにおいて同一尤度を持つ複数のパラメータセットがえられたとき、SR＝Mesnilとしてcheck.SRfitを呼び出し、gammaを変えながら尤度が一意に定まる値を探す関数}
\usage{
specify.Mesnil.gamma(resSR, n = 100, seed = 1, sigma = 5, gamma_ini = 10)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクトで、かつSR=Mesnilのもの}

\item{n}{初期値を変えてパラメータ推定する回数 (check.SRfitの引数)}

\item{seed}{\code{set.seed}で使用するseed(check.SRfitの引数)}

\item{sigma}{初期値を変えるときの生起乱数の標準偏差(check.SRfitの引数)}

\item{gamma_ini}{SR=Mesnilの曲がり具合を決めるパラメータ デフォルトは10で、9,8,…,2,1,0.9,0.8,…と複数のパラメータが得られるまで探していく。}
}
\description{
戻り値は一意のパラメータに収束した時のgamma
}
