% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{SRplot_gg}
\alias{SRplot_gg}
\title{再生産関係をプロットする関数}
\usage{
SRplot_gg(SR_result, refs = NULL, xscale = 1000,
  xlabel = "千トン", yscale = 1, ylabel = "尾",
  labeling.year = NULL, add.info = TRUE)
}
\arguments{
\item{SR_result}{fit.SRの結果のオブジェクト}
}
\description{
再生産関係をプロットする関数
}
