% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{SRregime_plot}
\alias{SRregime_plot}
\title{fit.SRregimeの結果で得られた再生産関係をプロットするための関数}
\usage{
SRregime_plot(SRregime_result, xscale = 1000, xlabel = "SSB",
  yscale = 1, ylabel = "R", labeling.year = NULL,
  show.legend = TRUE, legend.title = "Regime", regime.name = NULL,
  base_size = 16, add.info = TRUE)
}
\arguments{
\item{xscale}{X軸のスケーリングの値（親魚量をこの値で除す）}

\item{xlabel}{X軸のラベル}

\item{yscale}{Y軸のスケーリングの値（加入量をこの値で除す）}

\item{ylabel}{Y軸のラベル}

\item{labeling.year}{ラベルに使われる年}

\item{show.legend}{凡例を描くかどうか}

\item{legend.title}{凡例のタイトル（初期設定は\code{"Regime"}）}

\item{regime.name}{凡例に描かれる各レジームの名前（レジームの数だけ必要）}

\item{base_size}{\code{ggplot}のベースサイズ}

\item{add.info}{\code{AICc}や\code{regime.year}, \code{regime.key}などの情報を加えるかどうか}

\item{resSRregime}{\code{fit.SRregime}の結果}

\item{use.fit.SR}{パラメータの初期値を決めるのに\code{frasyr::fit.SR}を使う（時間の短縮になる）}
}
\description{
レジームごとの観察値と予測線が描かれる
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSRregime <- fit.SRregime(SRdata, SR="HS", method="L2", 
                            regime.year=c(1977,1989), regime.key=c(0,1,0),
                            regime.par = c("a","b","sd")[2:3])
g1 <- SRregime_plot(resSRregime, regime.name=c("Low","High"))
g1
}
}
