% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{test_sd0_future}
\alias{test_sd0_future}
\title{将来予測においてFが非常に小さい場合には決定論的予測と確率論的予測の平均がほぼ一致するかを確認するための関数}
\usage{
test_sd0_future(data_future, nsim = NULL, nyear = 10, future_range = NULL, ...)
}
\description{
HSを仮定していて、折れ点の下側から将来予測するような場合には、将来予測の最初のほうは一致しない。マッチングする年代などを工夫する必要がある
}
