% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{to_vpa_data}
\alias{to_vpa_data}
\title{tidy形式のVPAへのインプットデータをdata_dandlerへ渡す形式に変換する（暫定版）}
\usage{
to_vpa_data(x, label_name)
}
\description{
tidy形式のVPAへのインプットデータをdata_dandlerへ渡す形式に変換する（暫定版）
}
