% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{unlist_future_data}
\alias{unlist_future_data}
\title{ブートストラップ等でmake_future_dataをリスト化したものからfuture_vpaにインプットとして入れる将来予測のデータを生成する関数}
\usage{
unlist_future_data(data_future_list)
}
\arguments{
\item{data_future_list}{各要素にmake_future_dataのオブジェクトが格納されているリスト}
}
\description{
ブートストラップ等でmake_future_dataをリスト化したものからfuture_vpaにインプットとして入れる将来予測のデータを生成する関数
}
