% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rvpa_tmb.r
\encoding{UTF-8}
\name{use_rvpa_tmb}
\alias{use_rvpa_tmb}
\title{Ridge VPAの計算をTMBで実行するためにcppファイルのコンパイル等をする関数}
\usage{
use_rvpa_tmb(TmbFile = "rvpa_tmb", CppDir = system.file("executable",
  package = "frasyr"), RunDir = getwd())
}
\arguments{
\item{TmbFile}{Cppファイルの名前}

\item{CppDir}{Cppファイルが格納されているディレクトリ}

\item{RunDir}{実行するディレクトリ}
}
\description{
Ridge VPAの計算をTMBで実行するためにcppファイルのコンパイル等をする関数
}
\examples{
\dontrun{
use_rvpa_tmb()
}

}
