% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\name{validate_sr}
\alias{validate_sr}
\title{再生産関係のパラメタが想定した形になっているかを確認}
\usage{
validate_sr(SR = NULL, method = NULL, AR = NULL, res_sr = NULL)
}
\arguments{
\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker)}

\item{method}{最適化法（\code{"L2"}: 最小二乗法, \code{"L1"}: 最小絶対値法）}

\item{AR}{自己相関を推定するか(1), しないか(0)}

\item{res_sr}{fit.SR の結果}
}
\description{
再生産関係のパラメタが想定した形になっているかを確認
}
