% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa_hay.R
\encoding{UTF-8}
\name{vpa.hay}
\alias{vpa.hay}
\title{半期VPAで資源量推定を行う関数}
\usage{
vpa.hay(
  dat,
  con_tf = 3,
  con_fc = 3,
  rec = NULL,
  alpha = 1,
  tune = FALSE,
  abund = "B",
  min.age = 0,
  max.age = 0,
  ki = 1,
  link = "id",
  base = NA,
  af = NA,
  p.m = 0.5,
  stat.tf = "mean",
  index.w = NULL,
  use.index = "all",
  scale = 1000,
  sel.def = "max",
  b_est = TRUE,
  b_fix = 1,
  lambda = 0,
  no.est = FALSE,
  sel.const = FALSE,
  W_s = 1000,
  start_row = 1,
  ssb_type = 1,
  p_ssb = 0.25,
  p_slide = 1,
  est.method = "ml",
  optimizer = "nlm",
  Lower = -Inf,
  Upper = Inf,
  hessian = FALSE,
  dd = 10^(-10),
  wf = 5,
  AR = FALSE,
  Lo_AR = c(-Inf, -Inf, 0, -Inf),
  Up_AR = c(Inf, Inf, Inf, Inf),
  do_compile = FALSE,
  rho_init = 0,
  rho_est = TRUE,
  waa_new = NULL,
  rho_se = TRUE,
  cov_q = FALSE,
  X_q = NULL,
  rec_cor = NULL,
  eta = 1,
  p.init = 0.5
)
}
\description{
半期VPAで資源量推定を行う関数
}
\details{
ホッケ道北系群の資源評価を種に使われているコード。\code{frasyr::vpa}と対応していないコードもあったりするが、利用する人が適宜開発しながら、関数をアップデートしていくといいと思います。
}
