% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{yrs_used}
\alias{yrs_used}
\title{Extract years by keys}
\usage{
yrs_used(
  start,
  end,
  regime = use_regime_model,
  yr_switch = regime_year_MSY,
  key = regime_key_MSY,
  which_key = future_regime_MSY
)
}
\description{
Extract years by keys
}
