% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{beta.simulation}
\alias{beta.simulation}
\title{kobeII matrixの簡易版（Btarget, Blimitは決め打ちでβのみ変える)}
\usage{
beta.simulation(
  finput,
  beta_vector,
  year.lag = 0,
  type = "old",
  year_beta_change = NULL,
  output_type = "tidy"
)
}
\arguments{
\item{year_beta_change}{betaを変更する年の範囲。NULLの場合には全部の年を変える。}
}
\description{
kobeII matrixの簡易版（Btarget, Blimitは決め打ちでβのみ変える)
}
