% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{beta.simulation}
\alias{beta.simulation}
\title{kobeII matrixの簡易版（Btarget, Blimitは決め打ちでβのみ変える)}
\usage{
beta.simulation(finput, beta_vector, year.lag = 0)
}
\description{
kobeII matrixの簡易版（Btarget, Blimitは決め打ちでβのみ変える)
}
