% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{boo.vpa}
\alias{boo.vpa}
\title{bootstrapを実施する}
\usage{
boo.vpa(
  res,
  type = "index",
  B_ite = 5,
  method = "p",
  mean.correction = FALSE,
  B_cv = 0.2,
  ess = 200
)
}
\arguments{
\item{res}{vpaの出力値}

\item{type}{"index": 資源量指標値についてブートストラップ。"caa": 年齢別漁獲尾数についてブートストラップ}

\item{B_ite}{ブートストラップ法でのイタレーション数}

\item{method}{`type = "index"`の時に利用可。"p": パラメトリックブートストラップ。"n": ノンパラメトリックブートストラップ。"r": 残差をスムージングした後ブートストラップt法}

\item{mean.correction}{デフォルトは`FALSE`。対数正規分布の平均値の補正の有無。`type = "index"`と`type = "caa"`のいずれでも利用可能}

\item{B_cv}{`type = "caa"`の時に利用可。漁獲量の変動係数（i.e. 対数取った漁獲量の分散）。}

\item{ess}{`type = "caa"`の時に利用可。多項分布の有効サンプルサイズ。}
}
\description{
bootstrapを実施する
}
\seealso{
https://ichimomo.github.io/frasyr/articles/Diagnostics-for-VPA.html
}
