% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{boot.SR}
\alias{boot.SR}
\title{再生産関係のブートストラップ}
\usage{
boot.SR(Res, method = "p", n = 100, seed = 1)
}
\arguments{
\item{Res}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{method}{パラメトリック ("p") かノンパラメトリック ("n")}
}
\description{
①残差のパラメトリックブートストラップ、②残差のノンパラメトリックブートストラップ（リサンプリング）、③データのブートストラップ（リサンプリング）が行える
}
