% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{caa.est.mat}
\alias{caa.est.mat}
\title{年齢別生物パラメータとFと漁獲量を与えると与えた漁獲量と一致するFへの乗数を返す}
\usage{
caa.est.mat(naa, saa, waa, M, catch.obs, Pope, set_max1 = TRUE)
}
\description{
年齢別生物パラメータとFと漁獲量を与えると与えた漁獲量と一致するFへの乗数を返す
}
