% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{calc.rel.abund}
\alias{calc.rel.abund}
\title{年齢別パラメータを与えて、１年分前進計算する}
\usage{
calc.rel.abund(
  sel,
  Fr,
  na,
  M,
  waa,
  waa.catch = NULL,
  maa,
  min.age = 0,
  max.age = Inf,
  Pope = TRUE,
  ssb.coef = 0
)
}
\description{
年齢別パラメータを与えて、１年分前進計算する
}
