% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{calc_Fratio}
\alias{calc_Fratio}
\title{calculate F/Ftarget based on F_\%SPR multiplier}
\usage{
calc_Fratio(faa, waa, maa, M, SPRtarget = 30, waa.catch = NULL, Pope = TRUE)
}
\arguments{
\item{faa}{F at age}

\item{waa}{weight at age}

\item{maa}{maturity at age}

\item{M}{natural morality at age}

\item{SPRtarget}{target SPR}
}
\description{
calculate F/Ftarget based on F_\%SPR multiplier
}
