% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{calc_future_perSPR}
\alias{calc_future_perSPR}
\title{将来予測の結果オブジェクトから生物パラメータを取り出して、その生物パラメータをベースに、Fvectorで与えたF at ageに対応するSPRを計算して返す}
\usage{
calc_future_perSPR(
  fout = NULL,
  res_vpa = NULL,
  Fvector,
  Fmax = 10,
  max.age = Inf,
  target.col = NULL,
  target.year = NULL,
  SPRtarget = NULL,
  SPR_unit = "digit"
)
}
\arguments{
\item{fout}{将来予測のアウトプット（finputがない場合)。future_vpaの結果はformat_to_old_future関数をかまさないと動かない。}

\item{res_vpa}{Popeの式を使うかどうかだけのために利用。実際、漁獲量は計算していないので、不要といえば不要。}

\item{Fvector}{Fのベクトル}

\item{max.age}{SPR計算のときに考慮する最大年齢。デフォルトは無限大。}

\item{target.col}{将来予測の何列目の年を取り出すか（NULLの場合、最後の年）}

\item{target.year}{将来予測の何年目を取り出すか（年の名前）（NULLの場合、最後の年）}

\item{SPRtarget}{これを与えると、Fvectorを何倍すればここで指定した\%SPRと一致するかを返すようになる}
}
\description{
将来予測の結果オブジェクトから生物パラメータを取り出して、その生物パラメータをベースに、Fvectorで与えたF at ageに対応するSPRを計算して返す
}
