% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{calc_perspr}
\alias{calc_perspr}
\title{MSYを達成するときの\%SPRを計算する}
\usage{
calc_perspr(finput, Fvector, Fmax = 10, max.age = Inf,
  target.col = NULL)
}
\arguments{
\item{finput}{将来予測インプット}

\item{Fvector}{Fのベクトル}
}
\description{
MSYを達成するときの\%SPRを計算する
}
