% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{calc_steepness}
\alias{calc_steepness}
\title{Steepness (h) と関連するパラメータ (SB0,R0,B0)を計算する関数}
\usage{
calc_steepness(SR = "HS", rec_pars, M, waa, maa, plus_group = TRUE)
}
\arguments{
\item{SR}{"HS", "BH", "RI"のいずれか}

\item{rec_pars}{再生産関係のパラメータで\code{rec_pars$a},\code{rec_pars$b}で使用する}

\item{M}{年齢別自然死亡係数 (ベクトルで与えるか、年齢共通の場合\code{M=0.4}のようにしてもよい)}

\item{waa}{（親魚量の）年齢別体重}

\item{maa}{年齢別親魚量}

\item{plus_group}{最高齢がプラスグループかどうか}
}
\value{

}
\description{
Steepness (h) と関連するパラメータ (SB0,R0,B0)を計算する関数
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
               out.AR = TRUE)
rec_pars = resSR$pars
year <- "2017"
M = res_vpa$input$dat$M[,year]
waa = res_vpa$input$dat$waa[,year]
maa = res_vpa$input$dat$maa[,year]
Res_h = calc_steepness(SR=SR,rec_pars=rec_pars,M=M,waa=waa,maa=maa,plus_group=TRUE)
Res_h
}
}
