% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{calculate_all_pm}
\alias{calculate_all_pm}
\title{res_futureから考えられるほぼすべてのパフォーマンス指標をとりだし、2行のtibbleで返す}
\usage{
calculate_all_pm(res_future, SBtarget, SBlimit, SBban, SBmin)
}
\arguments{
\item{res_future}{future_vpaの返り値}

\item{SBtarget}{それを上回る・下回る確率を計算する。目標管理基準値}

\item{SBlimit}{それを上回る・下回る確率を計算する。限界管理基準値}

\item{SBban}{それを上回る・下回る確率を計算する。禁漁水準}

\item{SBmin}{それを上回る・下回る確率を計算する。過去最低親魚量}
}
\description{
res_futureから考えられるほぼすべてのパフォーマンス指標をとりだし、2行のtibbleで返す
}
\examples{
\dontrun{
  calculate_all_pm(res_future_0.8HCR, 0, 0, 0, 0)
}

}
