% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{calculate_all_pm}
\alias{calculate_all_pm}
\title{res_futureから考えられるほぼすべてのパフォーマンス指標をとりだし、2行のtibbleで返す}
\usage{
calculate_all_pm(
  res_future,
  SBtarget = -1,
  SBlimit = -1,
  SBban = -1,
  SBmin = -1,
  MSY = -1,
  is_scale = FALSE,
  unit = 1,
  period_extra = NULL,
  type = "AS",
  fun_period = mean
)
}
\arguments{
\item{res_future}{future_vpaの返り値}

\item{SBtarget}{それを上回る・下回る確率を計算する。目標管理基準値}

\item{SBlimit}{それを上回る・下回る確率を計算する。限界管理基準値}

\item{SBban}{それを上回る・下回る確率を計算する。禁漁水準}

\item{SBmin}{それを上回る・下回る確率を計算する。過去最低親魚量}

\item{MSY}{それを上回る・下回る確率を計算する。MSY}

\item{is_scale}{TRUEの場合、親魚量はSBtargetで、漁獲量はMSYで割った相対値が出力される。それ以外はそのまま}

\item{unit}{確率を出力するときの単位。100を入れると％単位で結果が返される}

\item{period_extra}{デフォルトではSSBminなどを一度でも下回るなど、期間を指定して計算する統計量はABC_year + 0:9, 0:4, 5:9, 1:10, 1:4, 6:10で決め打ちしているが、それ以外の期間を指定したいときにここの引数で与える}

\item{type}{"AS": age-structured from frasyr, "PM": production model from frapmr}
}
\description{
res_futureから考えられるほぼすべてのパフォーマンス指標をとりだし、2行のtibbleで返す
}
