% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{check.SRfit}
\alias{check.SRfit}
\title{再生産関係推定が収束しているかや最適解を得られているかを診断する関数}
\usage{
check.SRfit(
  resSR,
  n = 100,
  sigma = 5,
  seed = 1,
  output = FALSE,
  filename = "checkSRfit"
)
}
\arguments{
\item{resSR}{\code{fit.SR}か\code{fit.SRregime}のオブジェクト}

\item{n}{初期値を変えてパラメータ推定する回数}

\item{sigma}{初期値を変えるときの生起乱数の標準偏差}

\item{seed}{\code{set.seed}で使用するseed}

\item{output}{テキストファイルに結果を出力するか}

\item{filename}{ファイル名('.txt')がつく}
}
\description{
再生産関係推定が収束しているかや最適解を得られているかを診断する関数
}
