% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{derive_biopar}
\alias{derive_biopar}
\title{将来予測やVPAの結果から生物パラメータをとりだす}
\usage{
derive_biopar(res_obj = NULL, derive_year = NULL, stat = mean)
}
\arguments{
\item{res_obj}{VPAか将来予測の結果のオブジェクト。どちらでも良い。}

\item{derive_year}{生物パラメータとF at ageを取り出す期間（年の名前で指定）}

\item{stat}{取り出した期間のパラメータをここで指定する関数で処理する。基本は平均する（mean）

将来予測結果を入れる場合には複数のシミュレーション、年の間の結果をすべてstatする}
}
\description{
将来予測やVPAの結果から生物パラメータをとりだす
}
