% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_vpa.R
\encoding{UTF-8}
\name{do_jackknife_vpa}
\alias{do_jackknife_vpa}
\title{ジャックナイフ法を行う関数}
\usage{
do_jackknife_vpa(
  res,
  method = "index",
  what_plot = NULL,
  ncol = 5,
  plot_year = NULL
)
}
\arguments{
\item{res}{VPAの結果のオブジェクト}

\item{method}{データの抜き方。デフォルトは指数ごとに抜く("index")。他に各指数を各点ごとに抜く場合は"all"。}

\item{what_plot}{作図したい項目を選べる。引数を与えない場合（\code{NULL}）、全て（SSB, biomass, U, catch, Recruitment, fish_number, fishing_mortality, weight, maturity, catch_number）をプロットする。}

\item{ncol}{作図の列数。標準で5列なので、\code{what_plot}の数が5以下の場合は適宜変えた方がよい。}
}
\value{
返ってくる値:
    \code{JKplot_vpa} ジャックナイフ法を行った結果のプロット。
}
\description{
モデル診断法の一つです。影響力の強いデータや外れ値の検出に用いることができます。
}
\details{
method :
\code{"index"}を行うことで、どの指数シリーズの影響が大きいのか見ることができます（例：漁業CPUEより調査船調査の方が推定の影響が大きい）。
\code{"all"}を行うことで、どの指数シリーズのどの年の影響が大きいのか見ることができます。ただし組み合わせが多くなるので、結果が得られるまでに時間を要します。
}
\seealso{
作図について: \code{\link{plot_vpa}}
https://ichimomo.github.io/frasyr/doc/Diagnostics-for-VPA.html
}
\author{
濵邉昂平, 市野川桃子
}
