% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{draw_SRline}
\alias{draw_SRline}
\title{Draw fitted SR line(s) over observed stock-recruitment plots}
\usage{
draw_SRline(fitted, observed, show.year)
}
\arguments{
\item{fitted}{Fitted SR data created by fit.SR()}

\item{observed}{Observed data in format data.frame(year, SSB, R)}

\item{show.year}{If TRUE, show year label on each points}
}
\description{
Draw fitted SR line(s) over observed stock-recruitment plots
}
