% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{fit.SR}
\alias{fit.SR}
\title{再生産関係の推定}
\usage{
fit.SR(
  SRdata,
  SR = "HS",
  method = "L2",
  AR = 1,
  hessian = FALSE,
  w = rep(1, length(SRdata$R)),
  length = 20,
  max.ssb.pred = 1.3,
  p0 = NULL,
  out.AR = TRUE,
  rep.opt = FALSE
)
}
\arguments{
\item{SRdata}{\code{get.SRdata}で作成した再生産関係データ}

\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker)}

\item{method}{最適化法（\code{"L2"}: 最小二乗法, \code{"L1"}: 最小絶対値法）}

\item{AR}{自己相関を推定するか(1), しないか(0)}

\item{length}{初期値を決める際のgridの長さ}

\item{p0}{\code{optim}で設定する初期値}

\item{out.AR}{自己相関係数を一度再生産関係を推定したのちに、外部から推定するか（1), 内部で推定するか(0)}

\item{rep.opt}{\code{TRUE}で\code{optim}による最適化を収束するまで繰り返す}

\item{sd.obs}{L1のときにSDをRMSEに置き換えるか否か}
}
\value{
以下の要素からなるリスト
\describe{
\item{\code{input}}{使用した引数のリスト}
\item{\code{pars}}{推定されたパラメータ}
\item{\code{opt}}{\code{optim}の結果オブジェクト}
\item{\code{resid}}{再生産関係から予測値からの加入量の残差}
\item{\code{resid2}}{自己相関のを推定したうえでの加入の残差（自己相関なしの時\code{resid}と等しくなる)}
\item{\code{loglik}}{対数尤度}
\item{\code{k}}{推定したパラメータ数}
\item{\code{AIC}}{AIC (\code{out.AR=TRUE}のときは自己相関推定前の結果)}
\item{\code{AICc}}{AICc (\code{out.AR=TRUE}のときは自己相関推定前の結果)}
\item{\code{BIC}}{BIC (\code{out.AR=TRUE}のときは自己相関推定前の結果)}
\item{\code{AIC.ar}}{\code{out.AR=TRUE}のときに\code{acf}関数で得られた自己相関を推定しない場合(0)とする場合(1)のAICの差}
\item{\code{pred}}{予測された再生産関係}
}
}
\description{
3種類の再生産関係の推定を、最小二乗法か最小絶対値法で、さらに加入の残差の自己相関を考慮して行うことができる
}
\examples{
\dontrun{
data(res_vpa)
SRdata <- get.SRdata(res_vpa)
resSR <- fit.SR(SRdata, SR = c("HS","BH","RI")[1],
                method = c("L1","L2")[2], AR = 1,
                out.AR = TRUE, rep.opt = TRUE)
resSR$pars
}
}
