% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\name{fit.SR_tol}
\alias{fit.SR_tol}
\title{再生産関係のフィット（頑健版）}
\usage{
fit.SR_tol(..., n_check = 100, is_regime = FALSE, seed = 12345)
}
\arguments{
\item{...}{fit.SRへの引数}

\item{n_check}{check.SRfitに渡す引数n}

\item{seed}{check.SRfitに渡す引数seed。toolのほうで最初に12345にしていた関係で、デフォルトの引数は12345にしている}
}
\description{
fit.SRのwrapper。与えられたinputはそのままfit.SRに渡すが、計算したあとにcheck.SRfitを回し、大域解に達していない場合や同じ尤度を持つ複数の大域解がある場合、その中央値をとるres_SRを返す
}
