% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tmb.r
\name{future_vpa}
\alias{future_vpa}
\title{将来予測の実施関数}
\usage{
future_vpa(tmb_data, optim_method = "none", multi_init = 1,
  multi_lower = 0, multi_upper = 10, objective = "MSY",
  obj_value = 0, obj_stat = "mean", do_MSE = NULL,
  MSE_input_data = NULL, MSE_nsim = NULL, compile = FALSE,
  output_format = "new", attach_input = TRUE)
}
\arguments{
\item{tmb_data}{make_future_dataの返り値}
}
\description{
将来予測の実施関数
}
