% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r, R/future_tmb.r
\encoding{UTF-8}
\name{future_vpa}
\alias{future_vpa}
\title{将来予測の実施関数}
\usage{
future_vpa(
  tmb_data,
  optim_method = "none",
  multi_init = 1,
  multi_lower = 0,
  multi_upper = 10,
  objective = "MSY",
  obj_value = 0,
  obj_stat = "mean",
  do_MSE = NULL,
  MSE_input_data = NULL,
  MSE_nsim = NULL,
  compile = FALSE,
  output_format = "new",
  attach_input = TRUE
)

future_vpa(
  tmb_data,
  optim_method = "none",
  multi_init = 1,
  multi_lower = 0,
  multi_upper = 10,
  objective = "MSY",
  obj_value = 0,
  obj_stat = "mean",
  do_MSE = NULL,
  MSE_input_data = NULL,
  MSE_nsim = NULL,
  compile = FALSE,
  output_format = "new",
  attach_input = TRUE
)
}
\arguments{
\item{tmb_data}{make_future_dataの返り値}

\item{SPR_target}{目標とする\%SPR。NULL以外の値の場合、過去〜将来のそれぞれの年・シミュレーションが、目標とするF\%SPRに対して何倍にあたるか(F/Ftarget)を計算して、HCR_matの"Fratio"に入れる。HCRが生きている年については"beta_gamma"と一致するはず。}
}
\description{
将来予測の実施関数

将来予測の実施関数
}
