% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{future_vpa}
\alias{future_vpa}
\title{将来予測の実施関数}
\usage{
future_vpa(
  tmb_data,
  optim_method = "none",
  multi_init = 1,
  multi_lower = 1e-04,
  multi_upper = 10,
  objective = "MSY",
  obj_value = 0,
  obj_stat = "mean",
  do_MSE = NULL,
  MSE_input_data = NULL,
  MSE_nsim = NULL,
  MSE_sd = NULL,
  MSE_catch_exact_TAC = FALSE,
  compile = FALSE,
  output_format = "new",
  attach_input = TRUE,
  SPRtarget = NULL,
  calc_SPR_year_name = NULL
)
}
\arguments{
\item{tmb_data}{make_future_dataの返り値。将来の生物パラメータや再生産関係のシナリオを年齢×年×シミュレーション回数で指定した様々なarrayが含まれる。}

\item{optim_method}{"none"の場合、通常の将来予測を実施. "R": 以下のobj関係の設定とあわせてMSYなどを探索する. "tmb"もあるが、限定した設定でしか使えない}

\item{objective}{MSY:MSYの推定、PGY:PGYの値をobj_valueに入れる、percentB0:B0パーセント、何％にするかはobj_valueで指定, SSB:obj_valueで指定した特定の親魚資源量に一致するようにする}

\item{obj_value}{目的とする値}

\item{obj_stat}{目的関数を計算するときに利用する計算方法（"mean"だと平均、"median"だと中央値、"geomean"だと幾何平均）}

\item{do_MSE}{簡易MSEを実施するか}

\item{MSE_input_data}{簡易MSEを実施する場合、ABC計算するための将来予測を実施するための設定ファイル}

\item{MSE_nsim}{簡易MSEを実施する場合、ABC計算するための将来予測の繰り返し回数。ここを1にすると、決定論的な将来予測の漁獲量が用いられる。}

\item{MSE_sd}{簡易MSEをする場合の加入変動の大きさ。ここをゼロにすると決定論的な将来予測の値を将来の漁獲量として用いる。その場合MSE_nsimは自動的に２に設定される。単純なモデルの場合、ここがゼロでも多分問題ない。モデル平均を使っている場合にはちゃんとした簡易MSEをすること。}

\item{SPR_target}{目標とする\%SPR。NULL以外の値の場合、過去〜将来のそれぞれの年・シミュレーションが、目標とするF\%SPRに対して何倍にあたるか(F/Ftarget)を計算して、HCR_realizedの"Fratio"に入れる。HCRが生きている年については"beta_gamma"と一致するはず。}

\item{max_F}{漁獲量一定方策を実施する際のF at ageの最大値の上限（将来的にはmake_future_data関数に入れたい)}

\item{max_exploitation_rate}{漁獲量一定方策を実施する際のMを考慮した上での漁獲率の上限（将来的にはmake_future_data関数に入れたい)}
}
\description{
将来予測の実施関数
}
