% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\name{get.SPR}
\alias{get.SPR}
\title{VPA計算結果を入れると、毎年のF at ageがどのくらいのSPR, YPRに相当するかを返す}
\usage{
get.SPR(dres, target.SPR = 30, Fmax = 10, max.age = Inf)
}
\arguments{
\item{dres}{vpa関数の返り値}

\item{target.SPR}{目標とするSPR。この値を入れると、結果の$ysdataでその年のFが目標とするSPRを達成するためのFの何倍になっているかを返す。デフォルトは30が入っている}

\item{Fmax}{上の計算をするときに探索するFの乗数の最大値}

\item{max.age}{SPRやYPRの計算をするときに最大何歳まで考慮するか（デフォルトは無限大)}
}
\description{
VPA計算結果を入れると、毎年のF at ageがどのくらいのSPR, YPRに相当するかを返す
}
