% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.r
\encoding{UTF-8}
\name{get.SRdata}
\alias{get.SRdata}
\title{VPAの結果から再生産関係推定用のデータを作成する}
\usage{
get.SRdata(vpares, R.dat = NULL, SSB.dat = NULL,
  years = as.numeric(colnames(vpares$naa)), return.df = FALSE)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト}
}
\description{
VPAの結果から再生産関係推定用のデータを作成する
}
