% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tmb.r
\name{make_array}
\alias{make_array}
\title{将来予測用の三次元行列を与えられたら, pars.year間のパラメータを平均するか与えられたparをyear_replace_futureに入れる}
\usage{
make_array(d3_mat, pars, pars.year, year_replace_future)
}
\description{
将来予測用の三次元行列を与えられたら, pars.year間のパラメータを平均するか与えられたparをyear_replace_futureに入れる
}
