% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_tmb.r
\name{make_future_data}
\alias{make_future_data}
\title{future_vpaにインプットとして入れる将来予測の空のarrayを生成する関数}
\usage{
make_future_data(res_vpa, nsim = 1000, nyear = 50,
  future_initial_year_name = 2017, start_F_year_name = 2018,
  start_biopar_year_name = 2018, start_random_rec_year_name = 2018,
  waa_year, waa = NULL, waa_catch_year, waa_catch = NULL, maa_year,
  maa = NULL, M_year, M = NULL, faa_year, currentF = NULL,
  futureF = NULL, start_ABC_year_name = 2019, HCR_beta = 1,
  HCR_Blimit = -1, HCR_Bban = -1, HCR_year_lag = 0, res_SR = NULL,
  seed_number = 1, resid_type = "lognormal", resample_year_range = 0,
  bias_correction = TRUE, Pope = res_vpa$input$Pope)
}
\arguments{
\item{nsim}{シミュレーションの繰り返し回数}

\item{nyear}{将来予測の実施年数}

\item{future_initial_year_name}{将来予測の「初期値となる」年齢別資源尾数を参照する年。この年のFを使って翌年の個体群動態が将来予測で決定される}

\item{start_biopar_year_name}{将来の生物パラメータとして生物パラメータを置き換える年の最初の年}

\item{start_random_rec_year_name}{将来の加入を再生産関係から予測する最初の年}

\item{waa_year}{将来の年齢別体重を過去の平均値とする場合、過去のパラメータを平均する期間}

\item{waa}{将来の年齢別体重を直接与える場合}
}
\description{
future_vpaにインプットとして入れる将来予測の空のarrayを生成する関数
}
