% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r, R/future_tmb.r
\encoding{UTF-8}
\name{make_future_data}
\alias{make_future_data}
\title{future_vpaにインプットとして入れる将来予測の空のarrayを生成する関数}
\usage{
make_future_data(
  res_vpa,
  nsim = 1000,
  nyear = 50,
  plus_age = NULL,
  future_initial_year_name = 2017,
  start_F_year_name = 2018,
  start_biopar_year_name = 2018,
  start_random_rec_year_name = 2018,
  waa_year,
  waa = NULL,
  waa_catch_year,
  waa_catch = NULL,
  maa_year,
  maa = NULL,
  M_year,
  M = NULL,
  faa_year = NULL,
  currentF = NULL,
  futureF = NULL,
  start_ABC_year_name = 2019,
  HCR_beta = 1,
  HCR_Blimit = -1,
  HCR_Bban = -1,
  HCR_year_lag = 0,
  HCR_beta_year = NULL,
  Pope = res_vpa$input$Pope,
  fix_recruit = NULL,
  fix_wcatch = NULL,
  res_SR = NULL,
  seed_number = 1,
  resid_type = "lognormal",
  bias_correction = TRUE,
  resample_year_range = 0,
  backward_duration = 5,
  recruit_intercept = 0,
  model_average_option = NULL,
  regime_shift_option = NULL
)

make_future_data(
  res_vpa,
  nsim = 1000,
  nyear = 50,
  plus_age = NULL,
  future_initial_year_name = 2017,
  start_F_year_name = 2018,
  start_biopar_year_name = 2018,
  start_random_rec_year_name = 2018,
  waa_year,
  waa = NULL,
  waa_catch_year,
  waa_catch = NULL,
  maa_year,
  maa = NULL,
  M_year,
  M = NULL,
  faa_year = NULL,
  currentF = NULL,
  futureF = NULL,
  start_ABC_year_name = 2019,
  HCR_beta = 1,
  HCR_Blimit = -1,
  HCR_Bban = -1,
  HCR_year_lag = 0,
  HCR_beta_year = NULL,
  Pope = res_vpa$input$Pope,
  fix_recruit = NULL,
  fix_wcatch = NULL,
  res_SR = NULL,
  seed_number = 1,
  resid_type = "lognormal",
  bias_correction = TRUE,
  resample_year_range = 0,
  backward_duration = 5,
  recruit_intercept = 0,
  model_average_option = NULL,
  regime_shift_option = NULL
)
}
\arguments{
\item{res_vpa}{vpaの結果 (vpa関数の返り値)}

\item{nsim}{シミュレーションの繰り返し回数}

\item{nyear}{将来予測の実施年数}

\item{plus_age}{プラスグループの年齢。デフォルト値（NULL）ならfuture_initial_year_name年にNA以外の数値が入っている一番大きい年齢をプラスグループの年齢とする}

\item{future_initial_year_name}{将来予測の「初期値となる」年齢別資源尾数を参照する年。この年の年齢別資源尾数を使って翌年の個体群動態が将来予測で決定される}

\item{start_F_year_name}{将来予測でF全体にmultiplierを乗じる場合、multiplierを乗じる最初の年}

\item{start_biopar_year_name}{生物パラメータを将来の生物パラメータとして設定された値に置き換える年の最初の年}

\item{start_random_rec_year_name}{将来の加入を再生産関係から予測する最初の年}

\item{waa_year}{将来の年齢別体重を過去の平均値とする場合、過去のパラメータを平均する期間, maa_year, M_yearも同様}

\item{waa}{将来の年齢別体重を直接与える場合, maa, M_yearも同様}

\item{faa_year}{将来のFを過去の平均値とする場合、平均をとる年を指定する。下のcurrentF, futureFが指定されている場合にはこの設定は無視される}

\item{currentF}{start_ABC_yar_name以前に使うFのベクトル（いわゆるcurrent F）}

\item{futureF}{start_ABC_yar_name以降に使うFのベクトル（いわゆるFmsy）}

\item{start_ABC_year_name}{HCRを有効にする年}

\item{HCR_beta}{HCRのbeta}

\item{HCR_Bban}{HCRのBban}

\item{HCR_year_lag}{HCRするときにいつのタイミングのssbを参照するか.0の場合、ABC計算年のSSBを参照する。正の値1を入れると1年前のssbを参照する}

\item{HCR_beta_year}{betaを年によって変える場合。tibble(year=2020:2024, beta=c(1.3,1.2,1.1,1,0.9))　のようにtibble形式で与える}

\item{Pope}{漁獲方程式にPopeの近似式を使うかどうか。与えない場合には、VPAのオプションが引き継がれる}

\item{fix_recruit}{将来予測において再生産関係を無視して加入量を一定値で与える場合、その加入の値。list(year=2020, rec=1000)のように与える。}

\item{fix_wcatch}{将来予測において漁獲量をあらかじめ決める場合}

\item{res_SR}{再生産関係の推定関数 (fit.SR　of fit.SRregime) の返り値}

\item{seed_number}{乱数のシードの数}

\item{resid_type}{加入量の残差の発生方法。"lognormal":対数正規分布, "resample": リサンプリング, "backward": backward resampling}

\item{bias_correction}{将来予測でバイアス補正をするかどうか}

\item{resample_year_range}{"resampling", "backward"で有効。0の場合、推定に使ったデータから計算される残差を用いる。年の範囲を入れると、対象とした年の範囲で計算される残差を用いる。}

\item{backward_duration}{"backward"の場合、何年で１ブロックとするか。"backward"で有効。デフォルトは5 。}

\item{recruit_intercept}{将来の加入の切片。将来の加入は R=f(ssb) + intercept となる。}

\item{model_average_option}{model averagingをする場合のオプション. SR_matのlistとweightをlist形式で入れる(list(SR_list=list(res_SR1,res_SR2),weight=c(0.5,0.5)))}

\item{regime_shift_option}{res_SRにfit.SRregimeの返り値を入れた場合に指定する。将来予測で再生産関係のどのフェーズがおこるかを指定する。list(future_regime=将来のregimeの仮定。keyで指定された番号を入れる)}

\item{waa_fun}{log(weight)~log(number)の回帰式から将来のweightを予測する}

\item{start_waafun_year_name}{上記の設定がスタートする最初の年。それ以外の年は上で設定されたパラメータが使われる}

\item{special_setting}{list形式で与えるmake_future_dataの返り値のdataと同じ名前の要素について、最後にデータをここで示されたarrayのシミュレーション1回めの値で上書きする。arrayのデータに対してのみ有効。}

\item{HCR_Blimi}{HCRのBlimi}
}
\description{
future_vpaにインプットとして入れる将来予測の空のarrayを生成する関数

future_vpaにインプットとして入れる将来予測の空のarrayを生成する関数
}
