% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{make_kobeII_table}
\alias{make_kobeII_table}
\title{beta.simluationの結果などを読んで、kobeII talbeに整形する関数}
\usage{
make_kobeII_table(
  kobeII_data,
  res_vpa,
  year.catch = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssb = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbtarget = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssblimit = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbban = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbmin = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.risk = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  Bspecific = 0,
  Btarget = 0,
  Bmsy = NA,
  Blimit = 0,
  Bban = 0
)
}
\arguments{
\item{kobeII_data}{beta.simulationまたはconvert_future_list_tableの返り値}

\item{res_vpa}{VPAの結果}

\item{year.catch}{平均漁獲量の表を出力する期間。その他year.ssb(平均親魚量), year.ssbtarget(SBtargetを上回る確率)、、なども同様}
}
\description{
beta.simluationの結果などを読んで、kobeII talbeに整形する関数
}
\details{
tidy形式になっているkobeII_dataにおいて、HCR_name, betaの列のラベルの組み合わせを一つの管理方式として、その管理方式ごとに少尉予測の結果を集計する
}
