% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{make_kobeII_table}
\alias{make_kobeII_table}
\title{beta.simluationの結果などを読んで、kobeII talbeに整形する関数}
\usage{
make_kobeII_table(
  kobeII_data,
  res_vpa,
  year.catch = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssb = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.Fsakugen = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbtarget = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssblimit = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbban = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbmin = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.ssbmax = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.aav = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.risk = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  year.catchdiff = (max(as.numeric(colnames(res_vpa$naa))) + 1:10),
  Btarget = 0,
  Blimit = 0,
  Bban = 0
)
}
\arguments{
\item{kobeII_data}{beta.simulationの返り値}
}
\description{
beta.simluationの結果などを読んで、kobeII talbeに整形する関数
}
