% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_selection.R
\name{make_SR_dframe}
\alias{make_SR_dframe}
\title{Get fitted SR data with model names}
\usage{
make_SR_dframe(SR, method, vpares)
}
\arguments{
\item{SR}{再生産関係 (\code{"HS"}: Hockey-stick, \code{"BH"}: Beverton-Holt, \code{"RI"}: Ricker)}

\item{method}{最適化法（\code{"L2"}: 最小二乗法, \code{"L1"}: 最小絶対値法）}

\item{vpares}{VPAの結果のオブジェクト}
}
\description{
Get fitted SR data with model names
}
