% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_stock_table}
\alias{make_stock_table}
\title{Make table for stock assessment result}
\usage{
make_stock_table(
  result_vpa,
  result_msy,
  result_future,
  yr_future_start,
  yr_abc,
  faa_pre_abc,
  unit = "百トン"
)
}
\arguments{
\item{result_vpa}{Obejct returned from \code{vpa()}}

\item{result_msy}{Object MSY result created previous SC meeting...?}

\item{result_future}{Object returnd from \code{future_vpa()}}

\item{yr_future_start}{Year when future projection starts}
}
\description{
Make table for stock assessment result
}
