% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\name{out.vpa}
\alias{out.vpa}
\title{VPA結果をcsvファイルに出力する}
\usage{
out.vpa(res = NULL, srres = NULL, msyres = NULL,
  fres_current = NULL, fres_HCR = NULL, kobeII = NULL,
  filename = "vpa")
}
\arguments{
\item{res}{VPAの結果}

\item{srres}{fit.SRの結果}

\item{msyres}{est.MSYの結果}

\item{fres_current}{future.vpaの結果(Fcurrent)}

\item{fres_HCR}{future.vpaの結果(F with HCR)}

\item{kobeII}{kobeII.matrixの結果}
}
\description{
VPA結果をcsvファイルに出力する
}
