% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{out.vpa}
\alias{out.vpa}
\title{VPA結果をcsvファイルに出力する}
\usage{
out.vpa(res = NULL, srres = NULL, msyres = NULL,
  fres_current = NULL, fres_HCR = NULL, kobeII = NULL,
  filename = "vpa", csvname = NULL, pdfname = NULL,
  ci.future = c(0.1, 0.5, 0.9))
}
\arguments{
\item{res}{VPAの結果}

\item{srres}{fit.SRの結果}

\item{msyres}{est.MSYの結果}

\item{fres_current}{future.vpaの結果(Fcurrent)}

\item{fres_HCR}{future.vpaの結果(F with HCR)}

\item{kobeII}{kobeII.matrixの結果}

\item{filename}{csvファイルとpdfファイルの両方のファイル名を指定する場合（拡張子なしで指定）}

\item{csvname}{csvファイルのファイル名}

\item{pdfname}{pdfファイルのファイル名}
}
\description{
VPA結果をcsvファイルに出力する
}
