% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{plot_futures}
\alias{plot_futures}
\title{将来予測の複数の結果をggplotで重ね書きする}
\usage{
plot_futures(vpares, future.list = NULL,
  future.name = names(future.list), future_tibble = NULL,
  CI_range = c(0.1, 0.9), maxyear = NULL, font.size = 18, ncol = 3,
  is.plot.CIrange = TRUE, what.plot = c("Recruitment", "SSB",
  "biomass", "catch", "Fsakugen", "Fsakugen_ratio"), biomass.unit = 1,
  RP_name = c("Btarget", "Blimit", "Bban"), Btarget = 0, Blimit = 0,
  Bban = 0, MSY = 0, exclude.japanese.font = FALSE, n_example = 3,
  future.replicate = NULL, seed = 1)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト}

\item{future.list}{将来予測の結果をリストで並べたもの}

\item{n_example}{個々のシミュレーションの例を示す数}

\item{future.replicate}{どのreplicateを選ぶかを選択する。この場合n_exampleによる指定は無効になる}
}
\description{
将来予測の複数の結果をggplotで重ね書きする
}
