% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_HCR_by_catch}
\alias{plot_HCR_by_catch}
\title{縦軸が漁獲量のHCRを書く（traceの結果が必要）}
\usage{
plot_HCR_by_catch(
  trace,
  SBtarget,
  SBlim,
  SBban,
  Fmsy_vector,
  MSY,
  M_vector,
  biomass.unit = 1,
  beta = 0.8,
  col.multi2currf = "black",
  col.SBtarget = "#00533E",
  col.SBlim = "#edb918",
  col.SBban = "#C73C2E",
  col.Ftarget = "black",
  col.betaFtarget = "gray",
  is.text = TRUE,
  HCR_function_name = "HCR_default",
  Pope = TRUE,
  p.pope = 0.5,
  RP.label = c("目標管理基準値", "限界管理基準値", "禁漁水準")
)
}
\arguments{
\item{Fvector}{Fのベクトル}

\item{p.pope}{Popeの式でどこで漁獲するか（0.5=年の真ん中）}
}
\description{
縦軸が漁獲量のHCRを書く（traceの結果が必要）
}
