% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_HCR_by_catch}
\alias{plot_HCR_by_catch}
\title{縦軸が漁獲量のHCRを書く（traceの結果が必要）}
\usage{
plot_HCR_by_catch(
  trace,
  fout0.8,
  SBtarget,
  SBlim,
  SBban,
  Fmsy_vector,
  MSY,
  M_vector,
  biomass.unit = 1,
  beta = 0.8,
  col.multi2currf = "black",
  col.SBtarget = "#00533E",
  col.SBlim = "#edb918",
  col.SBban = "#C73C2E",
  col.Ftarget = "black",
  col.betaFtarget = "gray",
  is.text = TRUE,
  HCR_function_name = "HCR_default",
  Pope = TRUE,
  RP.label = c("\\u76ee\\u6a19\\u7ba1\\u7406\\u57fa\\u6e96\\u5024",
    "\\u9650\\u754c\\u7ba1\\u7406\\u57fa\\u6e96\\u5024", "\\u7981\\u6f01\\u6c34\\u6e96")
)
}
\arguments{
\item{fout}{将来予測のアウトプット（finputがない場合)}

\item{Fvector}{Fのベクトル}
}
\description{
縦軸が漁獲量のHCRを書く（traceの結果が必要）
}
