% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{plot_SRdata}
\alias{plot_SRdata}
\title{SRdataをプロットする}
\usage{
plot_SRdata(SRdata)
}
\arguments{
\item{vpares}{VPAの結果のオブジェクト}

\item{frag}{MSY計算時の将来予測で用いる引数のリスト}
}
\description{
SRdataをプロットする
}
