% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{plot_vpa}
\alias{plot_vpa}
\title{複数のVPAの結果を重ね書きする}
\usage{
plot_vpa(
  vpalist,
  vpatibble = NULL,
  what.plot = NULL,
  plot_year = NULL,
  legend.position = "top",
  vpaname = NULL,
  ncol = 2,
  scale_value = NULL
)
}
\arguments{
\item{vpalist}{vpaの返り値をリストにしたもの; 単独でも可}

\item{vpatibble}{tibble形式のVPA結果も可。この場合、convert_vpa_tibble関数の出力に準じる。複数のVPA結果がrbindされている場合は、列名"id"で区別する。}

\item{what.plot}{何の値をプロットするか. NULLの場合、全て（SSB, biomass, U, catch, Recruitment, fish_number, fishing_mortality, weight, maturity, catch_number）をプロットする。what.plot=c("SSB","Recruitment")とすると一部のみプロット。ここで指定された順番にプロットされる。}

\item{legend.position}{凡例の位置。"top" (上部), "bottom" (下部), "left" (左), "right" (右), "none" (なし)。}

\item{vpaname}{凡例につけるVPAのケースの名前。vpalistと同じ長さにしないとエラーとなる}

\item{ncol}{図の列数}
}
\description{
複数のVPAの結果を重ね書きする
}
\examples{
\dontrun{
data(res_vpa)
res_vpa2 <- res_vpa
res_vpa2$naa <- res_vpa2$naa*1.2

plot_vpa(list(res_vpa, res_vpa2), vpaname=c("True","Dummy"))
plot_vpa(list(res_vpa, res_vpa2), vpaname=c("True","Dummy"),
                 what.plot=c("SSB","fish_number"))

}

}
