% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\encoding{UTF-8}
\name{read.vpa}
\alias{read.vpa}
\title{csvファイルとしてまとめられた資源計算結果を読み込んでRのオブジェクトにする}
\usage{
read.vpa(
  tfile,
  caa.label = "catch at age",
  maa.label = "maturity at age",
  waa.label = "weight at age",
  waa.biomass.label = "weight at age for biomass calculation",
  waa.catch.label = "weight at age for catch calculation",
  M.label = "M at age",
  faa.label = "fishing mortality at age",
  Fc.label = "Current F",
  naa.label = "numbers at age",
  release.label = "release dat",
  Blimit = NULL,
  Pope = NULL,
  plus.group = NULL,
  fc.year = NULL
)
}
\arguments{
\item{tfile}{資源計算結果がまとめられたcsvファイルの名前}

\item{release.label}{放流データを読み込みたい場合。ここで指定したラベルをつけて、年齢を行、年を列とするデータを入れる（catch at ageと同じ構造だが、年齢は必要な年齢だけ取り出した形で良い）}

\item{Pope}{VPA計算時にどっちを使っているかここで設定する（TRUE or FALSE）。デフォルトはNULLで、その場合にはcaa,faa,naaの関係から自動判別するが、自動判別の結果が出力されるので、それをみて正しく判断されているか確認してください。}

\item{plus.group}{プラスグループを考慮するかどうか。こちらについても、NULLの場合にはfaaとnaaの関係から自動判別するが、結果を一応確認すること。}
}
\description{
csvファイルとしてまとめられた資源計算結果を読み込んでRのオブジェクトにする
}
