% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{read.vpa}
\alias{read.vpa}
\title{csvファイルとしてまとめられた資源計算結果を読み込んでRのオブジェクトにする}
\usage{
read.vpa(tfile, caa.label = "catch at age",
  maa.label = "maturity at age", waa.label = "weight at age",
  waa.biomass.label = "weight at age for biomass calculation",
  waa.catch.label = "weight at age for catch calculation",
  M.label = "M at age", faa.label = "fishing mortality at age",
  Fc.label = "Current F", naa.label = "numbers at age",
  Blimit = NULL, Pope = NULL, fc.year = NULL)
}
\arguments{
\item{tfile}{資源計算結果がまとめられたcsvファイルの名前}
}
\description{
csvファイルとしてまとめられた資源計算結果を読み込んでRのオブジェクトにする
}
