% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\name{ref.F}
\alias{ref.F}
\title{再生産関係を仮定しない管理基準値計算のための関数}
\usage{
ref.F(res, sel = NULL, waa = NULL, maa = NULL, M = NULL,
  waa.catch = NULL, M.year = NULL, waa.year = NULL,
  maa.year = NULL, rps.year = NULL, max.age = Inf, min.age = 0,
  d = 0.001, Fspr.init = 0.5, Fmax.init = 1.5, F0.1.init = 0.7,
  pSPR = seq(10, 90, by = 10), iterlim = 1000, plot = TRUE,
  Pope = FALSE, F.range = seq(from = 0, to = 2, length = 101))
}
\arguments{
\item{res}{VPAの出力結果}
}
\description{
再生産関係を仮定しない管理基準値計算のための関数
}
