% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{retro.est}
\alias{retro.est}
\title{レトロスペクティブ解析の実施}
\usage{
retro.est(res, n = 5, stat = "mean", init.est = FALSE,
  b.fix = TRUE)
}
\arguments{
\item{res}{VPAの出力}
}
\description{
レトロスペクティブ解析の実施
}
