% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r, R/future_tmb.r
\encoding{UTF-8}
\name{sample_backward}
\alias{sample_backward}
\title{過去にさかのぼってブロックサンプリングをおこなう}
\usage{
sample_backward(residual, n, duration)

sample_backward(residual, n, duration)
}
\arguments{
\item{residual}{リサンプリングする残差}

\item{n}{将来にわたって何年分のリサンプリング残差を作るか}

\item{duration}{1ブロックの年の長さ}
}
\description{
過去にさかのぼってブロックサンプリングをおこなう

過去にさかのぼってブロックサンプリングをおこなう
}
\examples{

set.seed(1)
res <- sample_backward(rep(1:5,each=5), 30, 5)
apply(matrix(res,5,6),2,min)


set.seed(1)
res <- sample_backward(rep(1:5,each=5), 30, 5)
apply(matrix(res,5,6),2,min)

}
