% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.r
\encoding{UTF-8}
\name{set_SR_mat}
\alias{set_SR_mat}
\title{将来予測用の再生産関係の設定を行う関数}
\usage{
set_SR_mat(
  res_vpa = NULL,
  start_random_rec_year_name,
  SR_mat,
  res_SR,
  seed_number,
  resid_type = "lognormal",
  bias_correction = TRUE,
  resample_year_range = NA,
  backward_duration = 5,
  recruit_intercept = 0,
  setting_release = NULL,
  recruit_age = 0,
  scale_ssb = 1,
  scale_R = 1,
  model_average_option = NULL,
  regime_shift_option = NULL,
  fix_recruit = NULL
)
}
\arguments{
\item{res_vpa}{VPAの推定結果}

\item{start_random_rec_year_name}{ランダム加入を仮定する最初の年}

\item{SR_mat}{将来予測用の再生産関係パラメータが格納する３次元行列}

\item{res_SR}{再生産関係の推定結果}

\item{seed_number}{シード番号}

\item{resid_type}{残差の発生パターン；対数正規分布は"lognormal"、単純リサンプリングは"resampling"、backward-resamplingは"backward"}

\item{resample_year_range}{"resampling", "backward"で有効。年の範囲を入れると、対象とした年の範囲で計算される残差を用いる。}

\item{backward_duration}{"backward"の場合、何年で１ブロックとするか。"backward"で有効。デフォルトは5 。}

\item{setting_release}{詳細はmake_future_dataへ}

\item{model_average_option}{model averagingをする場合のオプション. SR_matのlistとweightをlist形式で入れる(list(SR_list=list(res_SR1,res_SR2),weight=c(0.5,0.5))). 上で設定されたres_SRは使われない.}

\item{regime_shift_option}{レジームシフトを仮定する場合のオプション. この場合, res_SRにはfit.SRregimeの結果オブジェクトを入れる. オプションの設定は list(future_regime=将来のregimeの仮定。keyで指定された番号を入れる)}
}
\description{
再生産関係をres_SRで与えると、res_vpaを見ながら残差を再計算したのち、start_random_rec_year_name以降の加入のdeviationを計算しSR_mat[,,"deviance"]に入れる。
}
