% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\encoding{UTF-8}
\name{SRplot_gg}
\alias{SRplot_gg}
\title{再生産関係をプロットする関数}
\usage{
SRplot_gg(
  SR_result,
  refs = NULL,
  xscale = 1000,
  xlabel = "\\u5343繝医Φ",
  yscale = 1,
  ylabel = "\\u5c3e",
  labeling.year = NULL,
  add.info = TRUE,
  recruit_intercept = 0,
  plot_CI = FALSE,
  CI = 0.9
)
}
\arguments{
\item{SR_result}{fit.SRの結果のオブジェクト}

\item{refs}{管理基準値 (list(Blimit=0, Bmsy=10, Bban=0))}
}
\description{
再生産関係をプロットする関数
}
\examples{
\dontrun{
  data(res_sr_HSL1)
  plot_SR(res_sr_HSL1)
  plot_SR(res_sr_HSL1, refs=list(Blimit=20000, Bmsy=60000, Bban=0),
          recruit_intercept=100, plot_CI=TRUE)

  SRdata2 <- res_sr_HSL1$input$SRdata \%>\% as.data.frame()
  SRdata2$weight <- 1
  SRdata2$release <- 100 # 謾ｾ豬√ョ繝ｼ繧ｿ縺後≠繧句<a0>ｴ蜷<88>
  SRdata2$weight[length(SRdata2$weight)] <- 0 # 譛蠕後<81>ｮ蟷ｴ繧偵ヵ繧｣繝<83>繝医↓菴ｿ繧上↑縺<84>險ｭ螳<9a>
  res_SR2 <- fit.SR(SRdata=SRdata2, method="L1", AR=0)
  plot_SR(res_SR2)
}

}
