% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvpa.r
\encoding{UTF-8}
\name{vpa}
\alias{vpa}
\title{VPAによる資源計算を実施する}
\usage{
vpa(dat, sel.f = NULL, tf.year = 2008:2010, rec.new = NULL,
  rec = NULL, rec.year = 2010, rps.year = 2001:2010,
  fc.year = 2009:2011, last.year = NULL, last.catch.zero = FALSE,
  faa0 = NULL, naa0 = NULL, f.new = NULL, Pope = TRUE,
  p.pope = 0.5, tune = FALSE, abund = "B", min.age = 0,
  max.age = 0, link = "id", base = NA, af = NA, p.m = 0.5,
  omega = NULL, index.w = NULL, use.index = "all", scale = 1000,
  hessian = TRUE, alpha = 1, maxit = 5, d = 1e-04,
  min.caa = 0.001, plot = FALSE, plot.year = 1998:2015,
  term.F = "max", plus.group = TRUE, stat.tf = "mean",
  add.p.est = NULL, add.p.ini = NULL, sel.update = FALSE,
  sel.def = "max", max.dd = 1e-06, ti.scale = NULL, tf.mat = NULL,
  eq.tf.mean = FALSE, no.est = FALSE, est.method = "ls",
  b.est = FALSE, est.constraint = FALSE, q.const = 1:length(abund),
  b.const = 1:length(abund), q.fix = NULL, b.fix = NULL,
  sigma.const = 1:length(abund), fixed.index.var = NULL,
  max.iter = 100, optimizer = "nlm", Lower = -Inf, Upper = Inf,
  p.fix = NULL, lambda = 0, beta = 2, penalty = "p",
  ssb.def = "i", ssb.lag = 0, TMB = FALSE, sel.rank = NULL,
  p.init = 0.2, sigma.constraint = 1:length(abund), eta = NULL,
  eta.age = 0, tmb.file = "rvpa_tmb")
}
\arguments{
\item{dat}{data for vpa}

\item{sel.f}{最終年の選択率}

\item{tf.year}{terminal Fをどの年の平均にするか}

\item{rec.new}{翌年の加入を外から与える}

\item{rec}{rec.yearの加入}

\item{rec.year}{加入を代入する際の年}

\item{rps.year}{翌年のRPSをどの範囲の平均にするか}

\item{fc.year}{Fcurrentでどの範囲を参照するか}

\item{last.year}{vpaを計算する最終年を指定（retrospective analysis）}

\item{last.catch.zero}{TRUEなら強制的に最終年の漁獲量を0にする}

\item{faa0}{sel.update=TRUEのとき，初期値となるfaa}

\item{naa0}{sel.update=TRUEのとき，初期値となるnaa}

\item{f.new}{}

\item{Pope}{Popeの近似式を使うかどうか}

\item{p.pope}{Popeの式でどこで漁獲するか（0.5 = 真ん中の時期）}

\item{tune}{tuningをするかどうか}

\item{abund}{tuningの際，何の指標に対応するか}

\item{min.age}{tuning指標の年齢参照範囲の下限}

\item{max.age}{tuning指標の年齢参照範囲の上限}

\item{link}{tuningのlink関数}

\item{base}{link関数が"log"のとき，底を何にするか}

\item{af}{資源量指数が年の中央のとき，af=0なら漁期前，af=1なら漁期真ん中，af=2なら漁期後となる}

\item{p.m}{}

\item{index.w}{tuning indexの重み}

\item{use.index}{}

\item{scale}{重量のscaling}

\item{hessian}{}

\item{alpha}{最高齢と最高齢-1のFの比 F_a = alpha*F_{a-1}}

\item{maxit}{石岡・岸田/平松の方法の最大繰り返し数}

\item{d}{石岡・岸田/平松の方法の収束判定基準}

\item{min.caa}{caaに0があるとき，0をmin.caaで置き換える}

\item{plot}{tuningに使った資源量指数に対するフィットのプロット}

\item{plot.year}{上のプロットの参照年}

\item{term.F}{terminal Fの何を推定するか: "max" or "all"}

\item{plus.group}{}

\item{stat.tf}{最終年のFを推定する統計量（年齢別に与えること可）}

\item{add.p.est}{追加で最高齢以外のfaaを推定する際．年齢を指定する．}

\item{add.p.ini}{}

\item{sel.update}{チューニングVPAにおいて，選択率を更新しながら推定}

\item{sel.def}{sel.update=TRUEで選択率を更新していく際に，選択率をどのように計算するか．最大値を1とするか，平均値を1にするか...}

\item{max.dd}{sel.updateの際の収束判定基準}

\item{ti.scale}{資源量の係数と切片のscaling}

\item{tf.mat}{terminal Fの平均をとる年の設定．NA行列に平均をとる箇所に1を入れる．}

\item{eq.tf.mean}{terminal Fの平均値を過去のFの平均値と等しくする}

\item{no.est}{パラメータ推定しない．}

\item{est.method}{推定方法 （ls = 最小二乗法，ml = 最尤法）}

\item{b.est}{bを推定するかどうか}

\item{est.constraint}{制約付き推定をするかどうか}

\item{q.const}{qパラメータの制約（0は推定しないで1にfix）}

\item{b.const}{bパラメータの制約（0は推定しないで1にfix）}

\item{q.fix}{}

\item{b.fix}{}

\item{sigma.const}{sigmaパラメータの制約}

\item{fixed.index.var}{}

\item{max.iter}{q,b,sigma計算の際の最大繰り返し数}

\item{optimizer}{}

\item{Lower}{}

\item{Upper}{}

\item{p.fix}{}

\item{lambda}{ridge回帰係数}

\item{beta}{penaltyのexponent  (beta = 1: lasso, 2: ridge)}

\item{penalty}{}

\item{ssb.def}{i: 年はじめ，m: 年中央, l: 年最後}

\item{ssb.lag}{0: no lag, 1: lag 1}

\item{TMB}{TMBで高速計算する場合TMB=TRUE (事前にuse_rvpa_tmb()を実行)}

\item{p.init}{}

\item{eta}{Fのpenaltyを分けて与えるときにeta.ageで指定した年齢への相対的なpenalty (0~1)}

\item{eta.age}{Fのpenaltyを分けるときにetaを与える年齢(0 = 0歳（加入）,0:1 = 0~1歳)}

\item{tmb.file}{TMB=TRUEのとき使用するcppファイルの名前}
}
\description{
VPAによる資源計算を実施する
}
